% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allLM_pw.R
\name{allLM_pw}
\alias{allLM_pw}
\title{compute all pairwise linear models with optional transformations}
\usage{
allLM_pw(fmla, mae, xtx = force, ytx = force)
}
\arguments{
\item{fmla}{A formula specifying assays using Elist element names}

\item{mae}{A MultiAssayExperiment instance}

\item{xtx}{a function that transforms the independent variable}

\item{ytx}{a function that transforms the dependent variable}
}
\value{
a list with components 'mods' (lm results) and 'tslopes' (t stats for slope)
}
\description{
compute all pairwise linear models with optional transformations
}

