% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocMultiAssay-methods.R
\docType{methods}
\name{assay,ExpressionSet,ANY-method}
\alias{assay,ExpressionSet,ANY-method}
\title{Assay accessor}
\usage{
\S4method{assay}{ExpressionSet,ANY}(x)
}
\arguments{
\item{x}{An experiment object with data}
}
\value{
A basic representation of internal data
}
\description{
\code{assay} is used to obtain raw data
}

