% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-methods.R
\docType{methods}
\name{assay,matrix,ANY-method}
\alias{assay,matrix,ANY-method}
\title{Harmonize show to assay of a \code{matrix} object}
\usage{
\S4method{assay}{matrix,ANY}(x)
}
\arguments{
\item{x}{A \code{matrix} object}
}
\description{
Harmonize show to assay of a \code{matrix} object
}

