% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay-methods.R
\docType{methods}
\name{assay,RangedRaggedAssay,ANY-method}
\alias{assay,RangedRaggedAssay,ANY-method}
\title{Create a Matrix of score values using a GRanges or own ranges}
\usage{
\S4method{assay}{RangedRaggedAssay,ANY}(x, i = 1L, mcolname = "score",
  ranges = NULL, background = NULL, make.names = NULL)
}
\arguments{
\item{x}{A \linkS4class{RangedRaggedAssay} or \link{GRangesList} class}

\item{i}{Argument from generic (default 1L)}

\item{mcolname}{A single \code{character} string indicating the inner
metadata column name to use for creating a matrix (must indicate a numeric
variable)}

\item{ranges}{A \link{GRanges} class identifying the ranges of interest}

\item{background}{A single value for the non-matching background values in
the matrix (e.g., 2 for diploid genomes)}

\item{make.names}{logical (default FALSE) whether to automatically create
names from either the ranges argument (if available) or the
\code{RangedRaggedAssay} (e.g., "chr1:2-3:+")}
}
\value{
A \code{matrix} of values from the score column of the metadata.
}
\description{
This function can take a GRanges argument and use each range to check for
overlaps with any of the current ranges in the first argument and return a
score value from the corresponding metadata. This function will only operate
on fully disjoint ranges (see \code{isDisjoint} for details). It can only
work if metadata is present and there is a "score" column in the metadata.
Please see example on how to add metadata to a
\linkS4class{RangedRaggedAssay} or \link{GRangesList} class. This function
uses the \link{overlapsAny} function from the \code{GenomicRanges} package.
}
\examples{
example("RangedRaggedAssay")

## Add some phony metadata to the RangedRaggedAssay
metadata(myRRA) <- list(snparrray1 = DataFrame(score = 1),
snparray2 = DataFrame(score = 1),
snparray3 = DataFrame(score = 3))

assay(myRRA, background = 2)

}

