% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assayMatrix.R
\name{assayMatrix}
\alias{assayMatrix}
\title{Create a Matrix of score values using a GRanges or own ranges}
\usage{
assayMatrix(x, ranges = NULL, background = NA, use.names = FALSE)
}
\arguments{
\item{x}{A \linkS4class{RangedRaggedAssay} or \link{GRangesList} class}

\item{ranges}{A \link{GRanges} class identifying the ranges of interest}

\item{background}{A single value for the non-matching background values in
the matrix (e.g., 2 for diploid genomes)}

\item{use.names}{logical (default FALSE) whether to use names in the
given ranges argument or in the provided 'x' argument}
}
\value{
A \code{matrix} of values from the score column of the metadata.
}
\description{
This function can take a GRanges argument and use each range to check for
overlaps with any of the current ranges in the first argument and return a
score value from the corresponding metadata. This function will only operate
on fully disjoint ranges (see \code{isDisjoint} for details). It can only
work if metadata is present and there is a "score" column in the metadata.
Please see example on how to add metadata to a
\linkS4class{RangedRaggedAssay} or \link{GRangesList} class. This function
uses the \link{overlapsAny} function from the \code{GenomicRanges} package.
}
\examples{
example("RangedRaggedAssay")

## Add some phony metadata to the RangedRaggedAssay
metadata(myRRA) <- list(snparrray1 = DataFrame(score = 1),
snparray2 = DataFrame(score = 1),
snparray3 = DataFrame(score = 3))

assayMatrix(myRRA, background = 2)

}

