% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocMultiAssay.R
\docType{package}
\name{biocMultiAssay-package}
\alias{biocMultiAssay}
\alias{biocMultiAssay-package}
\title{MultiAssayExperiment: Build an integrative multiassay container}
\description{
MultiAssayExperiment allows the manipulation of related multiassay 
datasets with partially overlapping samples, associated metadata at 
the level of an entire study, and at the level of the "biological unit".
The biological unit may be a patient, plant, yeast strain, etc.
}
\details{
The package hierarch of information: 
\itemize{
\item study
\item experiments
\item samples
}
}

