% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSex.R
\name{clusterSex}
\alias{clusterSex}
\title{Check sex expression against clinical sex}
\usage{
clusterSex(MultiAssayExperiment, pDataCols, rownames, experiments, seed = 1L)
}
\arguments{
\item{MultiAssayExperiment}{A MultiAssayExperiment object}

\item{pDataCols}{Select columns from the MultiAssayExperiment pData
DataFrame}

\item{rownames}{Features to be used for clustering
(e.g., a set of gene names)}

\item{experiments}{A \code{character} vector indicating assays of interest
in the ExperimentList}

\item{seed}{A single integer value passed to \link{set.seed}}
}
\value{
A DataFrame with appended cluster and center values
}
\description{
Function that outputs a \link{DataFrame} with participant ID, sample ID,
the select pData column, the expression values for select rownames,
and the center values for each gene by cluster.
}
\examples{
example(MultiAssayExperiment)
clusterSex(myMultiAssayExperiment, pDataCols = "sex",
    rownames = c("XIST", "RPS4Y1", "KDM5D"),
    experiments = "RNASeqGene", seed = 42L)

}

