% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-methods.R
\docType{methods}
\name{collect}
\alias{collect}
\alias{collect,ANY-method}
\alias{collect,RangedRaggedAssay-method}
\alias{collect,ExperimentList-method}
\alias{collect,MultiAssayExperiment-method}
\title{Reshape raw data from an object}
\usage{
collect(object, ...)

\S4method{collect}{ANY}(object, ...)

\S4method{collect}{RangedRaggedAssay}(object, ...)

\S4method{collect}{ExperimentList}(object, ...)

\S4method{collect}{MultiAssayExperiment}(object, pDataCols = NULL, ...)
}
\arguments{
\item{object}{Any supported class object}

\item{...}{Additional arguments for the \link{RangedRaggedAssay}
\code{assay} method. See below.}

\item{pDataCols}{selected pData columns to include in the resulting output}
}
\value{
Tall and skinny \code{\linkS4class{DataFrame}}
}
\description{
The collect function collects data from the \code{\link{ExperimentList}}
in a \code{\link{MultiAssayExperiment}} and returns a uniform
\code{\link{DataFrame}}. The resulting DataFrame has columns indicating
primary, rowname, colname and value. This method can optionally include
pData columns with the \code{pDataCols} argument for a
\code{MultiAssayExperiment} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{ANY}: ANY class method, works with ExpressionSet and
SummarizedExperiment classes as well as matrix

\item \code{RangedRaggedAssay}: \linkS4class{RangedRaggedAssay} class method to return
matrix of selected \dQuote{mcolname} column, defaults to score

\item \code{ExperimentList}: Collect data from the \code{ExperimentList} class
returns list of DataFrames

\item \code{MultiAssayExperiment}: Overarching \code{MultiAssayExperiment} class method
returns a small and skinny DataFrame. The \code{pDataCols} arguments allows
the user to append pData columns to the long and skinny DataFrame.
}}
\examples{
example("RangedRaggedAssay")
collect(myRRA, background = 0)

}
\seealso{
\code{\link{assay,RangedRaggedAssay,missing-method}}
}

