% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocMultiAssay-methods.R
\docType{methods}
\name{colnames,ExpressionSet-method}
\alias{colnames,ExpressionSet-method}
\title{Colnames extractor}
\usage{
\S4method{colnames}{ExpressionSet}(x)
}
\arguments{
\item{x}{A compatible object with sample data}
}
\value{
Returns a vector of colnames or samplenames
}
\description{
\code{colnames} are used to obtain sample names from experiment data
}

