% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/downloadData.R
\name{downloadData}
\alias{downloadData}
\title{Download all the data from Amazon s3 bucket}
\usage{
downloadData(datadir)
}
\arguments{
\item{datadir}{The directory to where the S3 bucket will be dumped}
}
\value{
none
}
\description{
Creates directories and downloads respective files
}
\examples{
\dontrun{
downloadData("downloads")
}
}
\author{
Levi Waldron
}

