% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elist-class.R
\docType{methods}
\name{Elist,ANY-method}
\alias{Elist,ANY-method}
\title{Convert from \code{\link[S4Vectors]{SimpleList-class}}
to the multi-experiment data container}
\usage{
\S4method{Elist}{ANY}(x)
}
\arguments{
\item{x}{A \code{list} object}
}
\value{
An \code{\linkS4class{Elist}} class object
}
\description{
Convert from \code{\link[S4Vectors]{SimpleList-class}}
to the multi-experiment data container
}

