% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R, R/elist-class.R
\docType{methods}
\name{elist,MultiAssayExperiment-method}
\alias{elist,MultiAssayExperiment-method}
\alias{elist,SimpleList-method}
\alias{elist,list-method}
\alias{elist-class}
\title{An integrative container for assay data}
\usage{
\S4method{elist}{MultiAssayExperiment}(x)

\S4method{elist}{list}(x)

\S4method{elist}{SimpleList}(x)
}
\arguments{
\item{x}{A \code{list} object}
}
\value{
An \code{\linkS4class{elist}} class object
}
\description{
An integrative container for assay data

Convert to \code{\link[S4Vectors]{SimpleList-class}}
}
\section{Methods (by generic)}{
\itemize{
\item \code{elist}: 

\item \code{elist}: 

\item \code{elist}: 
}}

