% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\name{Elist<-}
\alias{Elist<-}
\title{Replace an \code{Elist} slot value with a given \code{Elist}
class object}
\usage{
Elist(x) <- value
}
\arguments{
\item{x}{A \code{MultiAssayExperiment} class object}

\item{value}{An \code{Elist} object to replace the existing
\code{Elist} slot}
}
\value{
A \code{Elist} class object
}
\description{
Replace an \code{Elist} slot value with a given \code{Elist}
class object
}
\examples{

\dontrun{
Elist(myMultiAssayExperiment) <- myNewElist
}

}

