% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elist-class.R
\name{Elist}
\alias{Elist}
\title{Generic Builder and Accessor Function}
\usage{
Elist(x = list())

Elist(x = list())
}
\arguments{
\item{x}{A \code{list} object}
}
\value{
An \code{\linkS4class{Elist}} class object
}
\description{
Generic Builder and Accessor Function

Convert to \code{\link[S4Vectors]{SimpleList-class}}
}

