% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentList-class.R
\name{ExperimentList-subset}
\alias{ExperimentList-subset}
\alias{[,ExperimentList,ANY,ANY,ANY-method}
\alias{[,ExperimentList,GRanges,ANY,ANY-method}
\alias{[[,ExperimentList,ANY,ANY-method}
\alias{[[,ExperimentList,ANY,ANY,ANY-method}
\alias{[[<-,ExperimentList,ANY,ANY-method}
\alias{[[<-,ExperimentList,ANY,ANY,ANY-method}
\alias{[<-,ExperimentList,ANY,ANY,ANY-method}
\title{Bracket subsetting operations for an ExperimentList object}
\usage{
\S4method{[}{ExperimentList,ANY,ANY,ANY}(x, i, j, k, ..., drop = TRUE)

\S4method{[}{ExperimentList,GRanges,ANY,ANY}(x, i, j, k, ..., drop = TRUE)

\S4method{[[}{ExperimentList,ANY,ANY}(x, i, j, ...)

\S4method{[[}{ExperimentList,ANY,ANY}(x, i, j, ...) <- value

\S4method{[}{ExperimentList,ANY,ANY,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{x}{A \code{MultiAssayExperiment} or \code{ExperimentList}}

\item{i}{Either a \code{character}, \code{integer}, \code{logical} or
\code{GRanges} object for subsetting by rows}

\item{j}{Either a \code{character}, \code{logical}, or \code{numeric} vector
for subsetting by \code{colData} rows. See details for more information.}

\item{k}{Either a \code{character}, \code{logical}, or \code{numeric} vector
for subsetting by assays}

\item{...}{Additional arguments passed on to lower level functions.}

\item{drop}{logical (default TRUE) whether to drop empty assay elements
in the \code{ExperimentList}}

\item{value}{An assay compatible with ExperimentList}
}
\description{
Single and double bracket extraction work similarly to
the equivalent methods in MultiAssayExperiment with the exception that
`sampleMap` information is not available for subsetting.
}
\examples{

example(MultiAssayExperiment)

## row subsetting for all assays
ExpList[1, ]

## column subsetting for a single assay
ExpList[, list(Affy = 1:2)]

## column subsetting for a single assay
ExpList[, , 1]

gr <- GRanges(c("chr2:11-12:+", "chr2:12-13:-"))
ExpList[gr, ]

## Extraction of the object
ExpList[[1L]]

}
\seealso{
subsetBy
}
