% Generated by roxygen2: do not edit by hand
% Please edit documentation in:
% R/ExperimentList-class.R
\name{ExperimentList}
\alias{ExperimentList}
\title{experiment Acessor function for the \code{ExperimentList} slot of a
\code{MultiAssayExperiment} object}
\usage{
ExperimentList(x)
}
\arguments{
\item{x}{A code{MultiAssayExperiment} class object}
}
\value{
A \code{ExperimentList} class object of experiment data
}
\description{
experiment Acessor function for the \code{ExperimentList} slot of a
\code{MultiAssayExperiment} object
}
\examples{
## Create an empty ExperimentList instance 
ExperimentList()

## Create array matrix and AnnotatedDataFrame to create an ExpressionSet class
arraydat <- matrix(seq(101, 108), ncol=4,
                   dimnames = list(
                     c("ENST00000294241", "ENST00000355076"), 
                     c("array1", "array2", "array3", "array4")
                   ))
arraypdat <- as(data.frame(
  slope53 = rnorm(4), 
  row.names = c("array1", "array2", "array3", "array4")), 
  "AnnotatedDataFrame")
exprdat <- Biobase::ExpressionSet(assayData=arraydat, phenoData=arraypdat)

## Create a sample methylation dataset
methyldat <- matrix(1:10, ncol = 5,
                    dimnames = list(
                      c("ENST00000355076", "ENST00000383706"),
                      c("methyl1", "methyl2", "methyl3", "methyl4", "methyl5")))

## Combine to a named list and call the ExperimentList constructor function
ExpList <- list(exprdat, methyldat)
names(ExpList) <- c("Affy", "Methyl450k")
myExperimentList <- ExperimentList(ExpList)
}

