% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\name{experiments<-}
\alias{experiments<-}
\title{Replace an \code{ExperimentList} slot value with a given
\code{ExperimentList} class object}
\usage{
experiments(object) <- value
}
\arguments{
\item{object}{A \code{MultiAssayExperiment} class object}

\item{value}{An \code{ExperimentList} object to replace the existing
\code{ExperimentList} slot}
}
\value{
A \code{ExperimentList} class object
}
\description{
Replace an \code{ExperimentList} slot value with a given
\code{ExperimentList} class object
}
\examples{
## Load a MultiAssayExperiment
example("MultiAssayExperiment")

## Replace with an empty ExperimentList
experiments(myMultiAssayExperiment) <- ExperimentList()

}
