% Generated by roxygen2: do not edit by hand
% Please edit documentation in:
% R/MultiAssayExperiment-class.R
\name{experiments}
\alias{experiments}
\title{Accessor function for the \code{ExperimentList} slot of a
\code{MultiAssayExperiment} object}
\usage{
experiments(x)
}
\arguments{
\item{x}{A \code{MultiAssayExperiment} class object}
}
\value{
A \code{ExperimentList} object of assay data
}
\description{
Accessor function for the \code{ExperimentList} slot of a
\code{MultiAssayExperiment} object
}
\examples{
example("MultiAssayExperiment")
experiments(myMultiAssayExperiment)
}

