% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-methods.R
\docType{methods}
\name{extract}
\alias{extract}
\alias{extract,ANY-method}
\alias{extract,ExpressionSet-method}
\alias{extract,MultiAssayExperiment-method}
\alias{extract,RangedRaggedAssay-method}
\alias{extract,SummarizedExperiment-method}
\title{Extract raw data from an object}
\usage{
extract(object)

\S4method{extract}{ExpressionSet}(object)

\S4method{extract}{ANY}(object)

\S4method{extract}{SummarizedExperiment}(object)

\S4method{extract}{RangedRaggedAssay}(object)

\S4method{extract}{MultiAssayExperiment}(object)
}
\arguments{
\item{object}{Any supported class object}
}
\value{
matrix class object
}
\description{
The [extract] function works to pull raw data from all the elements of
a supported class and to return a uniform data type, matrix.
}
\section{Methods (by class)}{
\itemize{
\item \code{ExpressionSet}: \code{\link{ExpressionSet}} class method

\item \code{ANY}: ANY class method, works best with matrices

\item \code{SummarizedExperiment}: \code{\link{SummarizedExperiment}} class
method

\item \code{RangedRaggedAssay}: \linkS4class{RangedRaggedAssay} class method to return
matrix of selected \dQuote{mcolname} column, defaults to score

\item \code{MultiAssayExperiment}: Overarching \code{MultiAssayExperiment} class method
returns list of matrices
}}
\examples{
example("RangedRaggedAssay")
extract(myRRA)

}

