% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biocMultiAssay-classes.R
\name{featExtractor}
\alias{featExtractor}
\title{Feature extractor for eSet, SummarizedExperiment, matrix, and GRangesList}
\usage{
featExtractor(x)
}
\arguments{
\item{x}{Either an \code{\linkS4class{ExpressionSet}}, \code{\linkS4class{GRangesList}}, \code{\linkS4class{SummarizedExperiment}} or \code{matrix} class object}
}
\value{
Returns either rownames or featureNames
}
\description{
Feature extractor for eSet, SummarizedExperiment, matrix, and GRangesList
}

