% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocMultiAssay-methods.R
\docType{methods}
\name{features}
\alias{features}
\alias{features,ExpressionSet-method}
\alias{features,GRanges-method}
\alias{features,GRangesList-method}
\alias{features,MultiAssayExperiment-method}
\alias{features,RangedSummarizedExperiment-method}
\alias{features,character-method}
\alias{features,matrix-method}
\title{Feature extractor method}
\usage{
features(x)

\S4method{features}{character}(x)

\S4method{features}{ExpressionSet}(x)

\S4method{features}{RangedSummarizedExperiment}(x)

\S4method{features}{matrix}(x)

\S4method{features}{GRanges}(x)

\S4method{features}{GRangesList}(x)

\S4method{features}{MultiAssayExperiment}(x)
}
\arguments{
\item{x}{Either an \code{\linkS4class{ExpressionSet}}, \code{\linkS4class{GRangesList}}, \code{\linkS4class{RangedSummarizedExperiment}} or \code{matrix} class object}
}
\value{
Returns either rownames or featureNames
}
\description{
Feature extractor method
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Return character if character

\item \code{ExpressionSet}: Get the featureNames for ExpressionSet

\item \code{RangedSummarizedExperiment}: Get a summary of rowRanges for RangedSummarizedExperiment

\item \code{matrix}: Get the rownames of a matrix

\item \code{GRanges}: Get the names of a GRanges

\item \code{GRangesList}: Get the summary of ranges for a GRangesList

\item \code{MultiAssayExperiment}: Get all the features for a MultiAssayExperiment
}}

