% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-methods.R
\docType{methods}
\name{gather}
\alias{gather}
\alias{gather,ExpressionSet-method}
\alias{gather,ANY-method}
\alias{gather,SummarizedExperiment-method}
\alias{gather,RangedRaggedAssay-method}
\alias{gather,MultiAssayExperiment-method}
\title{Reshape raw data from an object}
\usage{
gather(object)

\S4method{gather}{ExpressionSet}(object)

\S4method{gather}{ANY}(object)

\S4method{gather}{SummarizedExperiment}(object)

\S4method{gather}{RangedRaggedAssay}(object)

\S4method{gather}{MultiAssayExperiment}(object)
}
\arguments{
\item{object}{Any supported class object}
}
\value{
matrix class object
}
\description{
The gather function works to collect all data from the
\link{\code{ExperimentList}} class and to return a uniform
data type, matrix.
}
\section{Methods (by class)}{
\itemize{
\item \code{ExpressionSet}: \code{\link{ExpressionSet}} class method

\item \code{ANY}: ANY class method, works best with matrices

\item \code{SummarizedExperiment}: \code{\link{SummarizedExperiment}} class
method

\item \code{RangedRaggedAssay}: \linkS4class{RangedRaggedAssay} class method to return
matrix of selected \dQuote{mcolname} column, defaults to score

\item \code{MultiAssayExperiment}: Overarching \code{MultiAssayExperiment} class method
returns list of matrices
}}
\examples{
example("RangedRaggedAssay")
gather(myRRA)

}

