% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-methods.R
\name{gather_.default}
\alias{gather_.default}
\title{Reshape raw data from an object}
\usage{
gather_.default(object, ...)
}
\arguments{
\item{object}{Any supported class object}

\item{...}{Additional arguments for the \link{RangedRaggedAssay}
\code{assay} method. See below.}
}
\value{
Tall and skinny \code{\linkS4class{DataFrame}}
}
\description{
The gather function gathers data from the \code{\link{ExperimentList}}
in a \code{\link{MultiAssayExperiment}} and returns a uniform
\code{\link{DataFrame}}. The resulting DataFrame has columns indicating
primary, rowname, colname and value. This method can optionally include
pData columns with the \code{pDataCols} argument for a
\code{MultiAssayExperiment} object.
}
\examples{
example("RangedRaggedAssay")
gather(myRRA, background = 0)

}
\seealso{
\code{\link{assay,RangedRaggedAssay,missing-method}}
}

