% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocMultiAssay-methods.R
\docType{methods}
\name{getHits}
\alias{getHits}
\alias{getHits,ExpressionSet,character-method}
\alias{getHits,GRanges,GRanges-method}
\alias{getHits,GRangesList,GRanges-method}
\alias{getHits,MultiAssayExperiment,GRanges-method}
\alias{getHits,MultiAssayExperiment,character-method}
\alias{getHits,RangedSummarizedExperiment,GRanges-method}
\alias{getHits,matrix,character-method}
\title{Find hits by class type}
\usage{
getHits(subject, query, ...)

\S4method{getHits}{GRanges,GRanges}(subject, query, ...)

\S4method{getHits}{GRangesList,GRanges}(subject, query, ...)

\S4method{getHits}{RangedSummarizedExperiment,GRanges}(subject, query, ...)

\S4method{getHits}{ExpressionSet,character}(subject, query, ...)

\S4method{getHits}{matrix,character}(subject, query, ...)

\S4method{getHits}{MultiAssayExperiment,character}(subject, query, ...)

\S4method{getHits}{MultiAssayExperiment,GRanges}(subject, query, ...)
}
\arguments{
\item{subject}{Any valid element from the \code{\linkS4class{elist}} class}

\item{query}{Either a \code{character} vector or \code{\linkS4class{GRanges}} object used to search by name or ranges}

\item{...}{Additional arguments to findOverlaps}
}
\value{
Names of matched queries
}
\description{
Find hits by class type
}
\section{Methods (by class)}{
\itemize{
\item \code{subject = GRanges,query = GRanges}: Find overlaps and return names

\item \code{subject = GRangesList,query = GRanges}: Iteratively find overlaps and return names

\item \code{subject = RangedSummarizedExperiment,query = GRanges}: Find overlaps and return names for RangedSummarizedExperiment

\item \code{subject = ExpressionSet,query = character}: Find matching features in ExpressionSet

\item \code{subject = matrix,query = character}: Find matching features in matrix

\item \code{subject = MultiAssayExperiment,query = character}: Find all matching features by character

\item \code{subject = MultiAssayExperiment,query = GRanges}: Find all matching features by GRanges
}}

