% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocMultiAssay-methods.R
\docType{methods}
\name{getMap,Identify-method}
\alias{getMap,Identify-method}
\title{describeIn getMap Convert map from list to data.frame}
\usage{
\S4method{getMap}{Identify}(object)
}
\description{
describeIn getMap Convert map from list to data.frame
}

