% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocMultiAssay-methods.R
\docType{methods}
\name{getMap,MultiAssayView-method}
\alias{getMap,MultiAssayView-method}
\title{describeIn MultiAssayView Convert map from list to DataFrame}
\usage{
\S4method{getMap}{MultiAssayView}(object)
}
\description{
describeIn MultiAssayView Convert map from list to DataFrame
}

