% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocMultiAssay-methods.R
\name{getMap}
\alias{getMap}
\title{Convert Identify keeps slot to Map}
\usage{
getMap(object)
}
\arguments{
\item{object}{An \linkS4class{Identify} class object}
}
\value{
Returns a data.frame representation of samples
}
\description{
Convert Identify keeps slot to Map
}

