% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasAssay.R
\name{hasAssay}
\alias{hasAssay}
\title{Checking assay method for any class}
\usage{
hasAssay(object)
}
\arguments{
\item{object}{A \code{MultiAssayExperiment} or \code{named list}
object instance}
}
\value{
A \code{logical} value indicating method availability
}
\description{
The \code{hasAssay} function is intended for developers who would like to
include new classes into a \code{MultiAssayExperiment} instance.
It checks the methods tables of the \code{assay} function for the
specified class of the argument.
}
\examples{
lst <- structure(list(), .Names=character())
hasAssay(lst)

}
