% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-methods.R
\docType{methods}
\name{duplicated,MultiAssayExperiment-method}
\alias{duplicated,MultiAssayExperiment-method}
\alias{duplicated}
\alias{helpers}
\alias{intersectRows}
\alias{intersectPrimary}
\title{A group of helper functions for manipulating and cleaning a
MultiAssayExperiment}
\usage{
\S4method{duplicated}{MultiAssayExperiment}(x, incomparables = FALSE, ...)

intersectRows(object)

intersectPrimary(object)
}
\description{
A set of helper functions were created to help clean and
manipulate a MultiAssayExperiment object.

\itemize{
    \item complete.cases: Returns a logical vector corresponding to 'colData'
    rows that have data across all experiments
    \item duplicated: Returns a 'list' of 'LogicalList's that indicate
    what measurements originate from the same biological unit
    \item intersectRows: Takes all common rows across experiments,
    excludes experiments with empty rownames
    \item intersectPrimary: A wrapper for
    \link[=complete.cases,MultiAssayExperiment-method]{complete.cases} to
    return a MultiAssayExperiment with only those biological units that have
    measurements across all experiments
    \item mergeReplicates: A function that combines duplicated / repeated
    measurements across all experiments and is guided by the duplicated
    return value
}
}
