% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hnsc.R
\docType{data}
\name{hnsc}
\alias{hnsc}
\title{Head-Neck Squamous Cell Carcinoma (HNSC)}
\format{A \code{MultiAssayExperiment} with 2 experiments, each containing
400 rows and 20 columns:
\describe{
    \item{CNASNP}{Copy Number Alteration}
    \item{CNVSNP}{Copy Number Variation}
}}
\source{
The Cancer Genome Atlas
}
\usage{
hnsc
}
\description{
A \code{\link{MultiAssayExperiment}} object containing a random sample
of 20 patients with copy number data. For example purposes only.
}
\keyword{datasets}

