% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify-class.R
\docType{class}
\name{Identify-class}
\alias{Identify-class}
\title{An identifier class used for subsetting}
\description{
An identifier class used for subsetting
}
\section{Slots}{

\describe{
\item{\code{inassay}}{A \code{logical} indicator for a match by assay}

\item{\code{indim}}{A \code{list} indicating valid matches in each assay}

\item{\code{identifier}}{\code{ANY} object that in used to find matches}

\item{\code{drops}}{A \code{list} of excluded information due to subsetting}
}}

