% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/identifyByFeature.R
\name{identifyByFeature}
\alias{identifyByFeature}
\title{Identify by feature}
\usage{
identifyByFeature(MultiAssay, feature, requireall = FALSE)
}
\arguments{
\item{MultiAssay}{A \code{\link{MultiAssayExperiment}} object}

\item{feature}{A \code{character} vector containing feature names}
}
\value{
A logical list of matched assays
}
\description{
Identify by feature
}

