% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyBySample.R
\name{identifyBySample}
\alias{identifyBySample}
\title{Identify samples corresponding to row index in masterPheno
\code{identifyBySample} returns a \code{logical}}
\usage{
identifyBySample(MultiAssay, j)
}
\arguments{
\item{MultiAssay}{A \code{\linkS4class{MultiAssayExperiment}}}

\item{j}{A numeric vector referencing masterPheno data.frame row numbers}
}
\value{
A logical list of matched sample references
}
\description{
This function uses row indices of the masterPheno to return a logical list
}

