% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importcBioPortal.R
\name{importcBioPortal}
\alias{importcBioPortal}
\title{Title importcBioPortal: convert a .tar.gz file downloaded from
http://www.cbioportal.org/data_sets.jsp to a MultiAssayExperiment object}
\usage{
importcBioPortal(tgzfile, split.field = c("Tumor_Sample_Barcode", "ID"),
  names.field = c("Hugo_Symbol", "Entrez_Gene_Id"))
}
\arguments{
\item{tgzfile}{Path to the .tar.gz file as downloaded from cBioPortal web
page.}

\item{split.field}{A character vector of possible column names for the column
that is used to identify samples in a mutations or copy number file.}

\item{names.field}{A character vector of possible column names for the column
that is used to label ranges from a mutations or copy number file.}
}
\value{
A MultiAssayExperiment object
}
\description{
Title importcBioPortal: convert a .tar.gz file downloaded from
http://www.cbioportal.org/data_sets.jsp to a MultiAssayExperiment object
}
\examples{
# mae <- importcBioPortal("blca_mskcc_solit_2012.tar.gz") (not run)
}
\author{
Levi Waldron
}
