% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listToMap.R
\name{listToMap}
\alias{listToMap}
\title{Create a map of names from a list of names}
\usage{
listToMap(object, type = "colnames")
}
\arguments{
\item{object}{A \code{list} class object containing names of either 
experiments, assays or features.}

\item{type}{Any of the valid types of maps including colnames, rownames, 
and assays.}
}
\value{
A \code{DataFrame} class object of names
}
\description{
Takes a \code{list} of names and converts it to a 
\code{DataFrame} map depending on the type.
}
\examples{

\dontrun{
DFmap <- listToMap(listMap)
}

}

