% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapToList.R
\name{mapToList}
\alias{mapToList}
\title{Convert map from data.frame or DataFrame to list}
\usage{
mapToList(mapdf, assayCol = NULL)
}
\arguments{
\item{mapdf}{A \code{data.frame} or \code{DataFrame} object with
identifiers in the first column}

\item{assayCol}{A character vector of length one indicating the assay
names column}
}
\value{
A \code{list} object of DataFrames for each assay
}
\description{
\code{mapToList} returns a \code{list} of DataFrames
}

