% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\name{masterPheno}
\alias{masterPheno}
\title{Generic Accessor Functions}
\usage{
masterPheno(x)
}
\arguments{
\item{x}{A \code{\link{MultiAssayExperiment}} object}
}
\value{
A \code{data.frame} object
}
\description{
Generic Accessor Functions
}

