% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\docType{class}
\name{MatchedAssayExperiment-class}
\alias{MatchedAssayExperiment-class}
\title{An integrative and matched-samples class for experiment data}
\description{
This class supports the use of matched samples where an equal number
of observations are present in all assays.
}
