% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\docType{methods}
\name{metadata,MultiAssayExperiment-method}
\alias{metadata,MultiAssayExperiment-method}
\title{Generic Acessor Functions}
\usage{
\S4method{metadata}{MultiAssayExperiment}(x)
}
\arguments{
\item{x}{A \code{\link{MultiAssayExperiment}} object}
}
\value{
Any type of object describing the metadata
}
\description{
Generic Acessor Functions
}
\section{Methods (by class)}{
\itemize{
\item \code{MultiAssayExperiment}: 
}}

