% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R,
%   R/MultiAssayExperiment-methods.R
\docType{methods}
\name{MultiAssayExperiment-methods}
\alias{MultiAssayExperiment-methods}
\alias{experiments}
\alias{sampleMap}
\alias{experiments<-}
\alias{sampleMap<-}
\alias{sampleMap,MultiAssayExperiment-method}
\alias{experiments,MultiAssayExperiment-method}
\alias{pData,MultiAssayExperiment-method}
\alias{metadata,MultiAssayExperiment-method}
\alias{sampleMap<-,MultiAssayExperiment,DataFrame-method}
\alias{experiments<-,MultiAssayExperiment,ExperimentList-method}
\alias{pData<-,MultiAssayExperiment,DataFrame-method}
\alias{metadata<-,MultiAssayExperiment-method}
\alias{$<-,MultiAssayExperiment-method}
\alias{$,MultiAssayExperiment-method}
\alias{[[,MultiAssayExperiment,ANY,ANY-method}
\alias{[[<-,MultiAssayExperiment,ANY,ANY-method}
\title{Accessing/modifying slot information}
\usage{
\S4method{sampleMap}{MultiAssayExperiment}(x)

\S4method{experiments}{MultiAssayExperiment}(x)

\S4method{pData}{MultiAssayExperiment}(object)

\S4method{metadata}{MultiAssayExperiment}(x)

\S4method{sampleMap}{MultiAssayExperiment,DataFrame}(object) <- value

\S4method{experiments}{MultiAssayExperiment,ExperimentList}(object) <- value

\S4method{pData}{MultiAssayExperiment,DataFrame}(object) <- value

\S4method{metadata}{MultiAssayExperiment}(x, ...) <- value

\S4method{$}{MultiAssayExperiment}(x, name) <- value

\S4method{$}{MultiAssayExperiment}(x, name)

\S4method{[[}{MultiAssayExperiment,ANY,ANY}(x, i, j, ...)

\S4method{[[}{MultiAssayExperiment,ANY,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{x}{A \code{MultiAssayExperiment} object}

\item{object}{A \code{MultiAssayExperiment} object}

\item{value}{See details.}

\item{...}{Argument not in use}

\item{name}{A column in \code{pData}}

\item{i}{A \code{numeric} or \code{character} vector of length 1}

\item{j}{Argument not in use}
}
\value{
Accessors: Either a \code{sampleMap}, \code{ExperimentList}, or
\code{DataFrame} object

Setters: A \code{MultiAssayExperiment} object
}
\description{
A set of accessor and setter generic functions to extract
either the \code{sampleMap}, the \code{\link{ExperimentList}}, \code{pData},
or \code{metadata} slots of a \code{\link{MultiAssayExperiment}} object
}
\section{Accessors}{

Eponymous names for accessing \code{MultiAssayExperiment} slots with the
exception of the \link{ExperimentList} accessor named \code{experiments}.
\itemize{
   \item experiments: Access the \link{ExperimentList} slot
   \item [[: Access the \link{ExperimentList} slot
   \item $: Access a column in \code{pData}
}
}

\section{Setters}{

Setter method values (i.e., '\code{function(x) <- value}'):
\itemize{
    \item experiments<-: An \code{\link{ExperimentList}} object
    containing experiment data of supported classes
    \item sampleMap<-: A \code{\link{DataFrame}} object relating
    samples to biological units and assays
    \item pData<-: A \code{\link{DataFrame}} object describing the
    biological units
    \item metadata<-: A \code{list} object of metadata
    \item [[<-: Equivalent to the '\code{experiments<-}' setter method for
    convenience
    \item $<-: A vector to replace the indicated column in \code{pData}
}
}

\examples{
## Access the sampleMap
sampleMap(myMultiAssayExperiment)

## Replacement method for a MultiAssayExperiment sampleMap
sampleMap(myMultiAssayExperiment) <- DataFrame()

## Access the ExperimentList
experiments(myMultiAssayExperiment)

## Replace with an empty ExperimentList
experiments(myMultiAssayExperiment) <- ExperimentList()

## Access the metadata
metadata(myMultiAssayExperiment)

## Replace metadata with a list
metadata(myMultiAssayExperiment) <- list(runDate =
                                             format(Sys.time(), "\%B \%d, \%Y"))

## Access a column in pData
myMultiAssayExperiment$age

## Replace a column in pData
myMultiAssayExperiment$age <- myMultiAssayExperiment$age + 1
}
