% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-pkg.R
\docType{package}
\name{MultiAssayExperiment-package}
\alias{MultiAssayExperiment-package}
\title{MultiAssayExperiment: Build an integrative multi-assay container}
\description{
MultiAssayExperiment allows the manipulation of related multiassay
datasets with partially overlapping samples, associated metadata at
the level of an entire study, and at the level of the "biological unit".
The biological unit may be a patient, plant, yeast strain, etc.
}
\details{
The package hierarchy of information:
\itemize{
\item study
\item experiments
\item samples
}
}
\seealso{
Useful links:
\itemize{
  \item \url{http://waldronlab.io/MultiAssayExperiment/}
  \item Report bugs at \url{https://github.com/waldronlab/MultiAssayExperiment/issues}
}

}
\author{
\strong{Maintainer}: Marcel Ramos \email{marcel.ramos@sph.cuny.edu} (\href{https://orcid.org/0000-0002-3242-0582}{ORCID})

Authors:
\itemize{
  \item Marcel Ramos \email{marcel.ramos@sph.cuny.edu} (\href{https://orcid.org/0000-0002-3242-0582}{ORCID})
  \item Martin Morgan [contributor]
  \item Vincent J Carey [contributor]
  \item Levi Waldron \email{lwaldron.research@gmail.com}
}

Other contributors:
\itemize{
  \item Lori Shepherd [contributor]
  \item Hervé Pagès [contributor]
  \item MultiAssay SIG \email{biocmultiassay@googlegroups.com} [contributor]
  \item NCI (GrantNo.: U24CA289073) [funder]
}

}
