% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperimentToMAF.R
\name{MultiAssayExperimentToMAF}
\alias{MultiAssayExperimentToMAF}
\title{Convert MultiAssayExperiment to MAF class}
\usage{
MultiAssayExperimentToMAF(x, synAssay = "maf_syn", nonSynAssay = "maf_nonSyn")
}
\arguments{
\item{x}{A \code{MultiAssayExperiment} object}

\item{synAssay}{character(1) The name of the synonymous variant
classifications.}

\item{nonSynAssay}{character(1) The name of the assay with non-synonymous
variant classifications.}
}
\description{
Take a MultiAssayExperiment object with specific mutation
    assays and convert these into a maftools representation.
}
