% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayView-class.R, R/biocMultiAssay-methods.R
\docType{class}
\name{MultiAssayView-class}
\alias{.MultiAssayView}
\alias{MultiAssayView}
\alias{MultiAssayView-class}
\alias{colnames,MultiAssayView-method}
\alias{length,MultiAssayView-method}
\alias{names,MultiAssayView-method}
\alias{rownames,MultiAssayView-method}
\title{A class used for staging a subset operation}
\usage{
\S4method{rownames}{MultiAssayView}(x)

\S4method{colnames}{MultiAssayView}(x)

\S4method{names}{MultiAssayView}(x)

\S4method{length}{MultiAssayView}(x)
}
\description{
Use a \code{MultiAssayView} class to initialize a subsetting operation 
of a \code{\linkS4class{MultiAssayExperiment}} class object
}
\section{Methods (by generic)}{
\itemize{
\item \code{rownames}: Get a CharacterList of rownames

\item \code{colnames}: Get a CharacterList of colnames

\item \code{names}: Get a \code{character} vector of experiment names

\item \code{length}: Get the number of assays in the 
\code{MultiAssayExperiment} instance
}}
\section{Slots}{

\describe{
\item{\code{subject}}{A \code{\linkS4class{MultiAssayExperiment}} instance}

\item{\code{rowindex}}{An \code{IntegerList} indexing rows of subject
retained for subsequent operations}

\item{\code{colindex}}{An \code{IntegerList} indexing columns of sujbect
retained for subsequent operations}

\item{\code{assayindex}}{An \code{IntegerList} indexing assays of subject
retained for subsequent operations}
}}

