% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\name{pData}
\alias{pData}
\title{Accessor function for \code{pData} slot of a \code{MultiAssayExperiment}
object}
\usage{
pData(x)
}
\arguments{
\item{x}{A \linkS4class{MultiAssayExperiment} object}
}
\value{
A \code{DataFrame} object of phenotype/clinical data
}
\description{
Accessor function for \code{pData} slot of a \code{MultiAssayExperiment}
object
}

