% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwplot.R
\name{pwplot}
\alias{pwplot}
\title{plot features chosen from two assays in a MultiAssayExperiment}
\usage{
pwplot(fmla1, fmla2, mae, ytx = force, xtx = force, ...)
}
\arguments{
\item{fmla1}{defines sources of y and x using names of elist}

\item{fmla2}{defines scatterplot using names of features in the elist elements}

\item{mae}{the MultiAssayExperiment instance}

\item{ytx}{function transforming y before plot}

\item{xtx}{function transforming x before plot}

\item{\dots}{passed to plot}
}
\description{
plot features chosen from two assays in a MultiAssayExperiment
}
\details{
plot features from two assays
}

