% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangedRaggedAssay-class.R
\docType{class}
\name{RangedRaggedAssay-class}
\alias{RangedRaggedAssay,GRangesList-method}
\alias{RangedRaggedAssay-class}
\title{An extension of the GRangesList class}
\usage{
\S4method{RangedRaggedAssay}{GRangesList}(...)
}
\arguments{
\item{...}{A \code{GRangesList} object}
}
\value{
A \code{\linkS4class{RangedRaggedAssay}} class object
}
\description{
An extension of the GRangesList class

Convert GRangesList to RangedRaggedAssay
}
\section{Methods (by generic)}{
\itemize{
\item \code{RangedRaggedAssay}: Convert a GRangesList to a RangedRaggedAssay
}}

