% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-methods.R
\docType{methods}
\name{rearrange}
\alias{rearrange}
\alias{rearrange,ANY-method}
\alias{rearrange,RangedRaggedAssay-method}
\alias{rearrange,ExperimentList-method}
\alias{rearrange,MultiAssayExperiment-method}
\title{Reshape raw data from an object}
\usage{
rearrange(object, shape = "long", ...)

\S4method{rearrange}{ANY}(object, shape = "long", ...)

\S4method{rearrange}{RangedRaggedAssay}(object, shape = "long", ...)

\S4method{rearrange}{ExperimentList}(object, shape = "long", ...)

\S4method{rearrange}{MultiAssayExperiment}(object, shape = "long",
  colDataCols = NULL, ...)
}
\arguments{
\item{object}{Any supported class object}

\item{shape}{A single string indicating the shape of the resulting data,
options include \sQuote{long} and \sQuote{wide} (defaults to the former)}

\item{...}{Additional arguments for the \link{RangedRaggedAssay}
\code{assay} method. See below.}

\item{colDataCols}{selected colData columns to include in the resulting output}
}
\value{
Either a long or wide \code{\linkS4class{DataFrame}}
}
\description{
The rearrange function takes data from the \code{\link{ExperimentList}}
in a \code{\link{MultiAssayExperiment}} and returns a uniform
\code{\link{DataFrame}}. The resulting DataFrame has columns indicating
primary, rowname, colname and value. This method can optionally include
colData columns with the \code{colDataCols} argument for a
\code{MultiAssayExperiment} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{ANY}: ANY class method, works with classes such as
\link{ExpressionSet} and \link{SummarizedExperiment} as well as \code{matrix}

\item \code{RangedRaggedAssay}: \linkS4class{RangedRaggedAssay} class method to return
matrix of selected \dQuote{mcolname} column, defaults to score

\item \code{ExperimentList}: Rearrange data from the \code{ExperimentList} class
returns list of DataFrames

\item \code{MultiAssayExperiment}: Overarching \code{MultiAssayExperiment} class method
returns a small and skinny DataFrame. The \code{colDataCols} arguments allows
the user to append colData columns to the long and skinny DataFrame.
}}

\examples{
example("RangedRaggedAssay")
rearrange(myRRA, background = 0)

}
\seealso{
\code{\link{assay,RangedRaggedAssay,missing-method}}
}
