% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{reexports}
\alias{reexports}
\alias{DataFrame}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Click on the
function name to see their documentation.

\itemize{
    \item {
    S4Vectors: \linkS4class{DataFrame}
    }
}
}
\examples{
DataFrame()

}
