% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocMultiAssay-methods.R
\docType{methods}
\name{rownames,character-method}
\alias{rownames,character-method}
\title{Rownames extractor method}
\usage{
\S4method{rownames}{character}(x)
}
\arguments{
\item{x}{A compatible object with feature data}
}
\value{
Returns either rownames or featureNames
}
\description{
\code{rownames} are used to obtain feature names from experiment data
}

