% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocMultiAssay-methods.R
\docType{methods}
\name{rownames,ExpressionSet-method}
\alias{rownames,ExpressionSet-method}
\title{Harmonize featureNames to rownames of an \code{ExpressionSet} object}
\usage{
\S4method{rownames}{ExpressionSet}(x)
}
\arguments{
\item{x}{An \code{ExpressionSet} class object}
}
\value{
A \code{character} vector of row names
}
\description{
Harmonize featureNames to rownames of an \code{ExpressionSet} object
}

