% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-methods.R
\docType{methods}
\name{rownames,RangedSummarizedExperiment-method}
\alias{rownames,RangedSummarizedExperiment-method}
\title{Harmonize names of rowRanges to rownames of a
\code{RangedSummarizedExperiment} object}
\usage{
\S4method{rownames}{RangedSummarizedExperiment}(x)
}
\arguments{
\item{x}{A \code{RangedSummarizedExperiment} object}
}
\value{
A \code{character} vector of row names
}
\description{
Harmonize names of rowRanges to rownames of a
\code{RangedSummarizedExperiment} object
}

