% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biocMultiAssay-methods.R
\docType{methods}
\name{sampleExtractor}
\alias{sampleExtractor}
\alias{sampleExtractor,ExpressionSet-method}
\alias{sampleExtractor,GRangesList-method}
\alias{sampleExtractor,SummarizedExperiment-method}
\alias{sampleExtractor,matrix-method}
\title{Sample extractor generic}
\usage{
sampleExtractor(x)

\S4method{sampleExtractor}{ExpressionSet}(x)

\S4method{sampleExtractor}{SummarizedExperiment}(x)

\S4method{sampleExtractor}{matrix}(x)

\S4method{sampleExtractor}{GRangesList}(x)
}
\arguments{
\item{x}{Either an \code{\linkS4class{ExpressionSet}}, \code{\linkS4class{GRangesList}}, \code{\linkS4class{SummarizedExperiment}} or \code{matrix} class object}
}
\value{
Returns an object of the same class
}
\description{
Sample extractor generic
}
\section{Methods (by class)}{
\itemize{
\item \code{ExpressionSet}: 

\item \code{SummarizedExperiment}: 

\item \code{matrix}: 

\item \code{GRangesList}: 
}}

