% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\name{sampleMap<-}
\alias{sampleMap<-}
\title{Replace a slot value with a given \code{DataFrame}}
\usage{
sampleMap(object) <- value
}
\arguments{
\item{object}{A \code{MultiAssayExperiment} object}

\item{value}{A \code{DataFrame} object to replace the existing
\code{sampleMap}}
}
\value{
A \code{sampleMap} with replacement values
}
\description{
Replace a slot value with a given \code{DataFrame}
}
\examples{
## Load example
example("MultiAssayExperiment")

## Replacement method for a MultiAssayExperiment sampleMap
sampleMap(myMultiAssayExperiment) <- DataFrame()

}

