% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocMultiAssay-methods.R
\docType{methods}
\name{samples,ExpressionSet-method}
\alias{samples,ExpressionSet-method}
\alias{samples,GRangesList-method}
\alias{samples,MultiAssayExperiment-method}
\alias{samples,RangedSummarizedExperiment-method}
\alias{samples,matrix-method}
\title{Sample extractor generic}
\usage{
\S4method{samples}{ExpressionSet}(object)

\S4method{samples}{RangedSummarizedExperiment}(object)

\S4method{samples}{matrix}(object)

\S4method{samples}{GRangesList}(object)

\S4method{samples}{MultiAssayExperiment}(object)
}
\arguments{
\item{x}{Either an \code{\linkS4class{ExpressionSet}}, \code{\linkS4class{GRangesList}}, \code{\linkS4class{RangedSummarizedExperiment}} or \code{matrix} class object}
}
\value{
Returns an object of the same class
}
\description{
Sample extractor generic
}
\section{Methods (by class)}{
\itemize{
\item \code{ExpressionSet}: Get the sampleNames for ExpressionSet

\item \code{RangedSummarizedExperiment}: Get colnames for RangedSummarizedExperiment

\item \code{matrix}: Get the colnames of a matrix

\item \code{GRangesList}: Get the names of each list element for a GRangesList

\item \code{MultiAssayExperiment}: Get all the samples for a MultiAssayExperiment
}}

