% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveHDF5MultiAssayExperiment.R
\name{saveHDF5MultiAssayExperiment}
\alias{saveHDF5MultiAssayExperiment}
\title{Save a MultiAssayExperiment class object to HDF5 and Rds files}
\usage{
saveHDF5MultiAssayExperiment(
  x,
  dir = tempdir(),
  prefix = NULL,
  replace = FALSE,
  chunkdim = NULL,
  level = NULL,
  as.sparse = NA,
  verbose = NA
)
}
\description{
This function takes a `MultiAssayExperiment` object and uses the `assays`
functionality to create data matrices out of ti
}
\examples{

testDir <- tempdir()
saveHDF5MultiAssayExperiment(
    miniACC, dir = file.path(tempdir(), "test_mae"), replace = TRUE
)


}
