% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elist-class.R
\docType{methods}
\name{show,elist-method}
\alias{show,elist-method}
\title{Show method for \code{\linkS4class{elist}} class}
\usage{
\S4method{show}{elist}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{elist}} object.}
}
\value{
Returns a summary of contents for the \code{\linkS4class{elist}} class
exportMethod show
}
\description{
Show method for \code{\linkS4class{elist}} class
}

