% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\docType{methods}
\name{show,MultiAssayExperiment-method}
\alias{show,MultiAssayExperiment-method}
\title{Show method for \code{\linkS4class{MultiAssayExperiment}} class}
\usage{
\S4method{show}{MultiAssayExperiment}(object)
}
\arguments{
\item{object}{A \code{\link{MultiAssayExperiment}} object.}
}
\value{
Returns a summary of contents for
the \code{\link{MultiAssayExperiment}} class.
}
\description{
Show method for \code{\linkS4class{MultiAssayExperiment}} class
}

