% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayView-class.R
\docType{methods}
\name{show,MultiAssayView-method}
\alias{show,MultiAssayView-method}
\title{Show method for \code{\linkS4class{MultiAssayView}} class}
\usage{
\S4method{show}{MultiAssayView}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{MultiAssayView}} class object}
}
\value{
A summary of \code{\linkS4class{MultiAssayView}} class contents
}
\description{
Show method for \code{\linkS4class{MultiAssayView}} class
}

