% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage-class.R
\docType{methods}
\name{show,Stage-method}
\alias{show,Stage-method}
\title{Show method for \code{\linkS4class{Stage}} class}
\usage{
\S4method{show}{Stage}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{Stage}} class object}
}
\value{
A summary of \code{\linkS4class{Stage}} class contents
}
\description{
Show method for \code{\linkS4class{Stage}} class
}

