% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-class.R
\docType{class}
\name{stage-class}
\alias{stage-class}
\title{An identifier class used for staging a subset operation}
\description{
An identifier class used for staging a subset operation
}
\section{Slots}{

\describe{
\item{\code{query}}{Any class indicator needed to subset}

\item{\code{keeps}}{A \code{list} indicating valid matches in each assay}

\item{\code{drops}}{A \code{list} of excluded information due to subsetting}

\item{\code{type}}{A \code{character} vector indicating method used to search}
}}

