% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocMultiAssay-methods.R
\docType{methods}
\name{[,RangedRaggedAssay,ANY-method}
\alias{[,RangedRaggedAssay,ANY-method}
\title{Subset RangedRaggedAssay}
\usage{
\S4method{[}{RangedRaggedAssay,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A \code{\link{RangedRaggedAssay}} class}

\item{i}{Either a \code{character} or \code{GRanges} class object 
to subset by rows}

\item{j}{Either a \code{character}, \code{numeric}, or \code{logical} 
type for selecting columns (\code{\link[GenomicRanges]{GRangesList}} method)}

\item{...}{Any additional arguments passed on to subsetByOverlaps}
}
\value{
A \code{\link{RangedRaggedAssay}} class object
}
\description{
Subsetting a RangedRaggedAssay can be done using either rownames and column
names
}
\seealso{
\code{\link[IRanges]{findOverlaps-methods}}
}

