% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subset.R
\docType{methods}
\name{subset,MultiAssayExperiment-method}
\alias{subset,MultiAssayExperiment-method}
\title{Subset MultiAssayExperiment object
\code{subset} returns a subsetted \code{\linkS4class{MultiAssayExperiment}} object}
\usage{
\S4method{subset}{MultiAssayExperiment}(x, indicator, by, ...)
}
\arguments{
\item{indicator}{A \code{logical} vector or object of class \code{Identify} to use for subsetting}

\item{by}{A \code{character} vector of length one designating to subset either by samples, features, or assays}

\item{MultiAssay}{A \code{\linkS4class{MultiAssayExperiment}} object}
}
\description{
Subset MultiAssayExperiment object
\code{subset} returns a subsetted \code{\linkS4class{MultiAssayExperiment}} object
}

