% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-methods.R
\docType{methods}
\name{subsetByColData}
\alias{subsetByColData}
\alias{subsetByColData,MultiAssayExperiment,ANY-method}
\alias{subsetByColData,MultiAssayExperiment,character-method}
\title{Subset \code{MultiAssayExperiment} object by \code{colData} rows}
\usage{
subsetByColData(x, y)

\S4method{subsetByColData}{MultiAssayExperiment,ANY}(x, y)

\S4method{subsetByColData}{MultiAssayExperiment,character}(x, y)
}
\arguments{
\item{x}{A \code{MultiAssayExperiment} object}

\item{y}{Either a \code{numeric}, \code{character} or
\code{logical} object indicating what \code{colData} rows to select}
}
\value{
A \code{\link{MultiAssayExperiment}} object
}
\description{
Select biological units in a \code{MultiAssayExperiment} with
\code{subsetByColData}
}
\section{Methods (by class)}{
\itemize{
\item \code{x = MultiAssayExperiment,y = ANY}: Either a \code{numeric}, \code{character}, or
\code{logical} vector to apply a column subset of a
\code{MultiAssayExperiment} object

\item \code{x = MultiAssayExperiment,y = character}: Use a \code{character}
vector for subsetting column names
}}

\examples{
## Load a MultiAssayExperiment example
example("MultiAssayExperiment")

## Subset by character vector (Jack)
subsetByColData(myMultiAssayExperiment, "Jack")

## Subset by numeric index of colData rows (Jack and Bob)
subsetByColData(myMultiAssayExperiment, c(1, 3))

## Subset by logical indicator of colData rows (Jack and Jill)
subsetByColData(myMultiAssayExperiment, c(TRUE, TRUE, FALSE, FALSE))

}
