% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subsetByFeature.R
\name{subsetByFeature}
\alias{subsetByFeature}
\title{Subset MultiAssayExperiment object by Feature}
\usage{
subsetByFeature(MultiAssay, feature, ...)
}
\arguments{
\item{MultiAssay}{A \code{\link{MultiAssayExperiment}} object}

\item{feature}{A \code{character} vector containing feature names}
}
\value{
A \code{\link{MultiAssayExperiment}} object
}
\description{
Subset MultiAssayExperiment class by provided feature names
}

