% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-methods.R
\docType{methods}
\name{subsetBypData}
\alias{subsetBypData}
\alias{subsetBypData,MultiAssayExperiment,ANY-method}
\alias{subsetBypData,MultiAssayExperiment,character-method}
\title{Subset \code{MultiAssayExperiment} object by \code{pData} rows}
\usage{
subsetBypData(x, y)

\S4method{subsetBypData}{MultiAssayExperiment,ANY}(x, y)

\S4method{subsetBypData}{MultiAssayExperiment,character}(x, y)
}
\arguments{
\item{x}{A \code{MultiAssayExperiment} object}

\item{y}{Either a \code{numeric}, \code{character} or
\code{logical} object indicating what \code{pData} rows to select}
}
\value{
A \code{\link{MultiAssayExperiment}} object
}
\description{
Select biological units in a \code{MultiAssayExperiment} with
\code{subsetBypData}
}
\section{Methods (by class)}{
\itemize{
\item \code{x = MultiAssayExperiment,y = ANY}: Either a \code{numeric}, \code{character}, or
\code{logical} vector to apply a column subset of a
\code{MultiAssayExperiment} object

\item \code{x = MultiAssayExperiment,y = character}: Use a \code{character}
vector for subsetting column names
}}

\examples{
## Load a MultiAssayExperiment example
example("MultiAssayExperiment")

## Subset by character vector (Jack)
subsetBypData(myMultiAssayExperiment, "Jack")

## Subset by numeric index of pData rows (Jack and Bob)
subsetBypData(myMultiAssayExperiment, c(1, 3))

## Subset by logical indicator of pData rows (Jack and Jill)
subsetBypData(myMultiAssayExperiment, c(TRUE, TRUE, FALSE, FALSE))

}
