% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetByRow.R
\name{subsetByRow}
\alias{subsetByRow}
\title{Subset MultiAssayExperiment object by Feature}
\usage{
subsetByRow(MultiAssayExperiment, rowIndicator, drop = FALSE, ...)
}
\arguments{
\item{MultiAssayExperiment}{A \code{\link{MultiAssayExperiment}} object}

\item{rowIndicator}{Either a \code{character} vector or \code{\linkS4class{MultiAssayView}} class object containing feature names}

\item{drop}{logical indicating whether to coerce to lowest possible dimension after subsetting}

\item{...}{Additional arguments to pass to low level subsetting function}
}
\value{
A \code{\link{MultiAssayExperiment}} object
}
\description{
Subset MultiAssayExperiment class by provided feature names
}

