% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetBySample.R
\name{subsetBySample}
\alias{subsetBySample}
\title{Subset MultiAssayExperiment object
\code{subsetBySample} returns a subsetted \code{\linkS4class{MultiAssayExperiment}} object}
\usage{
subsetBySample(MultiAssay, identifier)
}
\arguments{
\item{MultiAssay}{A \code{\link{MultiAssayExperiment}} object}

\item{identifier}{An \linkS4class{stage} class object to be used for subsetting}
}
\description{
Subset MultiAssayExperiment object
\code{subsetBySample} returns a subsetted \code{\linkS4class{MultiAssayExperiment}} object
}

