% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocMultiAssay-methods.R
\docType{methods}
\name{subsetSample}
\alias{subsetSample}
\alias{subsetSample,ExpressionSet-method}
\alias{subsetSample,RaggedRangedAssay-method}
\alias{subsetSample,RangedSummarizedExperiment-method}
\alias{subsetSample,matrix-method}
\title{Subset by Sample generic}
\usage{
subsetSample(x, j, ...)

\S4method{subsetSample}{matrix}(x, j)

\S4method{subsetSample}{ExpressionSet}(x, j)

\S4method{subsetSample}{RangedSummarizedExperiment}(x, j)

\S4method{subsetSample}{RaggedRangedAssay}(x, j)
}
\arguments{
\item{x}{Either an \code{\linkS4class{ExpressionSet}}, \code{\linkS4class{RaggedRangedAssay}}, \code{\linkS4class{RangedSummarizedExperiment}} or \code{matrix} class object}

\item{j}{Either a \code{numeric}, \code{character}, or \code{logical} vector class for subsetting}

\item{...}{Additional arguments to pass}
}
\value{
Returns a subsetted \code{\linkS4class{MultiAssayExperiment}} object
}
\description{
Subset by Sample generic
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: Select columns of a matrix

\item \code{ExpressionSet}: Select colnames of an ExpressionSet

\item \code{RangedSummarizedExperiment}: Select column data of a RangedSummarizedExperiment

\item \code{RaggedRangedAssay}: Select colnames for a RaggedRangedAssay
}}

