% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TCGA_ov_v2rnaseq.R
\docType{data}
\name{TCGA_ov_v2rnaseq}
\alias{TCGA_ov_v2rnaseq}
\title{RNASeqV2 Ovarian Cancer data from TCGA}
\format{A \code{\link[GenomicRanges]{SummarizedExperiment-class}} object}
\source{
The National Cancer Institute, The Cancer Genome Atlas
}
\usage{
data(TCGA_ov_v2rnaseq)
}
\description{
Level 3 TCGA data
}

