% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/toListMap.R
\name{toListMap}
\alias{toListMap}
\title{Convert data.frame map to list map
\code{toListMap} returns a \code{list}}
\usage{
toListMap(dfmap)
}
\arguments{
\item{dfmap}{A data frame object with identifiers in the first column}
}
\description{
Convert data.frame map to list map
\code{toListMap} returns a \code{list}
}

