% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocMultiAssay-methods.R
\name{type}
\alias{type}
\title{Generic Accessor Functions}
\usage{
type(object)
}
\arguments{
\item{x}{A \code{\linkS4class{Stage}} class object}
}
\value{
A \code{character} atomic vector
}
\description{
Generic Accessor Functions
}

