% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maeUpset.R
\name{upset_samples}
\alias{upset_samples}
\title{create a generalized venn diagram analog for sample membership in multiple assays, using the upset algorithm in UpSetR}
\usage{
upset_samples(mae, nsets = length(experiments(mae)), nintersects = 24,
  order = "freq", idclip = function(x) substr(x, 1, 12), ...)
}
\arguments{
\item{mae}{instance of \code{\link[MultiAssayExperiment]{MultiAssayExperiment-class}}}

\item{nsets}{integer number of sets to analyze, as in \code{\link[UpSetR]{upset}}}

\item{nintersects}{integer number of intersections to display}

\item{order}{character, either 'freq' or 'degree' to indicate order of intersection size display, as in \code{\link[UpSetR]{upset}}}

\item{idclip}{a function that operates on \code{colnames(mae)}, to remove potentially assay-specific token components; use \code{force} if no clipping is needed}

\item{\dots}{parameters passed to \code{\link[UpSetR]{upset}}}
}
\description{
create a generalized venn diagram analog for sample membership in multiple assays, using the upset algorithm in UpSetR
}
\examples{
example(MultiAssayExperiment)
upset_samples(myMultiAssayExperiment, idclip=function(x) gsub("[a-z]", "", x))

}

