% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetSamples.R
\name{upsetSamples}
\alias{upsetSamples}
\title{Create a generalized Venn Diagram analog for sample membership in multiple
assays, using the upset algorithm in \code{UpSetR}}
\usage{
upsetSamples(MultiAssayExperiment, nsets = length(MultiAssayExperiment),
  nintersects = 24, order.by = "freq", ...)
}
\arguments{
\item{MultiAssayExperiment}{instance of
\code{\link[MultiAssayExperiment]{MultiAssayExperiment-class}}}

\item{nsets}{integer number of sets to analyze}

\item{nintersects}{Number of intersections to plot. If set to NA, all intersections will be plotted.}

\item{order.by}{How the intersections in the matrix should be ordered by. Options include frequency (entered as "freq"), degree, or both in any order.}

\item{...}{parameters passed to \code{\link[UpSetR]{upset}}}
}
\value{
Produces a visualization of set intersections using the UpSet matrix
design
}
\description{
Create a generalized Venn Diagram analog for sample membership in multiple
assays, using the upset algorithm in \code{UpSetR}
}
\examples{
data(miniACC)
upsetSamples(miniACC)
}
\author{
Vincent J Carey
}
