% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{API}
\alias{API}
\title{Refer to the API documentation}
\usage{
API(website = TRUE, shiny = FALSE)
}
\arguments{
\item{website}{(logical default TRUE) launch the website for the API}

\item{shiny}{(logical default FALSE) whether to launch the shiny version
of the API (experimental)}
}
\value{
Documentation via the GitHub wiki
}
\description{
\code{API} opens a browser to the API documentation
}
\examples{
## Runnable example does nothing

API(website = FALSE)

}
\author{
Vincent J Carey
}

