% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/API.R
\name{API}
\alias{API}
\title{Refer to the API documentation
\code{API} opens a browser to the API documentation}
\usage{
API()
}
\description{
Refer to the API documentation
\code{API} opens a browser to the API documentation
}

