% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{API}
\alias{API}
\title{Refer to the API documentation}
\usage{
API(shiny = FALSE)
}
\arguments{
\item{shiny}{(logical default FALSE) whether to launch the shiny version
of the API (experimental)}
}
\value{
Documentation via the GitHub wiki
}
\description{
\code{API} opens a browser to the API documentation
}
\examples{
## This will open in a browser window
\dontrun{
API()
}

}
\author{
Vincent J Carey
}

