% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocMultiAssay-methods.R
\docType{methods}
\name{assay,ExpressionSet,ANY-method}
\alias{assay,ExpressionSet,ANY-method}
\title{Harmonize exprs to assay of an \code{ExpressionSet} object}
\usage{
\S4method{assay}{ExpressionSet,ANY}(x)
}
\arguments{
\item{x}{An \code{ExpressionSet} object}
}
\description{
Harmonize exprs to assay of an \code{ExpressionSet} object
}

