\name{coerce-methods}
\docType{methods}
\alias{coerce}
\alias{coerce,ExpressionSet,SummarizedExperiment-method}
\title{ convert ExpressionSet to SummarizedExperiment }
\description{
convert ExpressionSet to SummarizedExperiment
}
\section{Methods}{
\describe{

\item{\code{signature(from = "ExpressionSet", to = "SummarizedExperiment")}}{
There are many ways of establishing mappable feature names in an
ExpressionSet instance.  This tool strives for reasonable generality.
}

}}
\keyword{methods}
