% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocMultiAssay-methods.R
\docType{methods}
\name{colnames,ExpressionSet-method}
\alias{colnames,ExpressionSet-method}
\title{Harmonize sampleNames to colnames of an \code{ExpressionSet} object}
\usage{
\S4method{colnames}{ExpressionSet}(x)
}
\arguments{
\item{x}{An \code{ExpressionSet} object}
}
\description{
Harmonize sampleNames to colnames of an \code{ExpressionSet} object
}

