% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elist-class.R
\docType{class}
\name{elist-class}
\alias{elist,SimpleList-method}
\alias{elist,list-method}
\alias{elist-class}
\title{An integrative container for assay data}
\usage{
\S4method{elist}{list}(x)

\S4method{elist}{SimpleList}(x)
}
\arguments{
\item{x}{A \code{list} object}
}
\value{
An \code{\linkS4class{elist}} class object
}
\description{
An integrative container for assay data

Convert to \code{\link[S4Vectors]{SimpleList-class}}
}
\section{Methods (by generic)}{
\itemize{
\item \code{elist}: Convert a list to a SimpleList to an elist

\item \code{elist}: Convert a SimpleList to an elist
}}

