% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elist-class.R, R/MultiAssayExperiment-methods.R, R/assay-methods.R
\docType{class}
\name{Elist-class}
\alias{.Elist}
\alias{Elist,ANY-method}
\alias{Elist,missing-method}
\alias{Elist-class}
\alias{assay,ANY,ANY-method}
\alias{assay,Elist,ANY-method}
\alias{colnames,Elist-method}
\alias{rownames,Elist-method}
\alias{show,Elist-method}
\title{A container for multi-experiment data}
\usage{
\S4method{Elist}{ANY}(x)

\S4method{Elist}{missing}(x)

\S4method{show}{Elist}(object)

\S4method{rownames}{Elist}(x)

\S4method{colnames}{Elist}(x)

\S4method{assay}{ANY,ANY}(x)

\S4method{assay}{Elist,ANY}(x)
}
\arguments{
\item{x}{A \code{list} object}

\item{object}{An \code{\linkS4class{Elist}} class object}
}
\value{
An \code{Elist} class object
}
\description{
The \code{Elist} class is a container that builds on 
the \code{SimpleList} with additional 
checks for consistency in experiment names and length.
It contains a \code{SimpleList} of experiments with sample identifiers.
One element present per experiment performed.
}
\details{
Convert from \code{SimpleList} or \code{list}
to the multi-experiment data container
}
\section{Methods (by generic)}{
\itemize{
\item \code{Elist}: Create an \code{Elist} object from an "ANY" class object, 
mainly \code{list}

\item \code{Elist}: Create an empty Elist for signature "missing"

\item \code{show}: Show method for \code{\linkS4class{Elist}} class

\item \code{rownames}: Get all the rownames of an \code{Elist}

\item \code{colnames}: Get sample names from an \code{Elist} object

\item \code{assay}: Get the assay data for the default ANY class

\item \code{assay}: Get the assay data from each element in the \link{Elist}
}}
\examples{
Elist()

}

