% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elist-class.R
\docType{class}
\name{Elist-class}
\alias{.Elist}
\alias{Elist-class}
\title{A container for multi-experiment data}
\description{
The \code{Elist} class is a container that builds on 
the \code{\link[S4Vectors]{SimpleList-class}} with additional 
checks for consistency in experiment names and length. It contains a \code{SimpleList}
of experiments with sample identifiers. One element present per experiment performed.
}

