% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elist-class.R
\docType{class}
\name{Elist-class}
\alias{.Elist}
\alias{Elist,ANY-method}
\alias{Elist,missing-method}
\alias{Elist-class}
\alias{show,Elist-method}
\title{A container for multi-experiment data}
\usage{
\S4method{Elist}{ANY}(x)

\S4method{Elist}{missing}(x)

\S4method{show}{Elist}(object)
}
\arguments{
\item{x}{A \code{list} object}

\item{object}{An \code{\linkS4class{Elist}} class object}
}
\value{
An \code{Elist} class object
}
\description{
The \code{Elist} class is a container that builds on 
the \code{SimpleList} with additional 
checks for consistency in experiment names and length.
It contains a \code{SimpleList} of experiments with sample identifiers.
One element present per experiment performed.
}
\details{
Convert from \code{SimpleList} or \code{list}
to the multi-experiment data container
}
\section{Methods (by generic)}{
\itemize{
\item \code{Elist}: Create an \code{Elist} object from an "ANY" class object, 
mainly \code{list}

\item \code{Elist}: Create an empty Elist for signature "missing"

\item \code{show}: Show method for \code{\linkS4class{Elist}} class
}}
\examples{
Elist()

}

