% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\name{Elist<-}
\alias{Elist<-}
\title{Replace an \code{Elist} slot value with a given \code{\linkS4class{Elist}}
class object}
\usage{
Elist(x) <- value
}
\arguments{
\item{x}{A \code{MultiAssayExperiment} class object}

\item{value}{An \code{\linkS4class{Elist}} object to replace the existing
\code{\linkS4class{Elist}} slot}
}
\description{
Replace an \code{Elist} slot value with a given \code{\linkS4class{Elist}}
class object
}

