% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\name{Elist<-}
\alias{Elist<-}
\title{Replace an \code{Elist} slot value with a given \code{Elist}
class object}
\usage{
Elist(object) <- value
}
\arguments{
\item{object}{A \code{MultiAssayExperiment} class object}

\item{value}{An \code{Elist} object to replace the existing
\code{Elist} slot}
}
\value{
A \code{Elist} class object
}
\description{
Replace an \code{Elist} slot value with a given \code{Elist}
class object
}
\examples{
## Load a MultiAssayExperiment
example("MultiAssayExperiment")

## Replace with an empty Elist
Elist(myMultiAssayExperiment) <- Elist()

}

