% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elist-class.R, R/MultiAssayExperiment-class.R
\name{Elist}
\alias{Elist}
\title{Convert from \code{\link[S4Vectors]{SimpleList-class}}
to the multi-experiment data container}
\usage{
Elist(x)

Elist(x)
}
\arguments{
\item{x}{A \code{list} object}

\item{x}{A \code{\link{MultiAssayExperiment}} object.}
}
\value{
An \code{\linkS4class{Elist}} class object

A \code{\link[S4Vectors]{SimpleList-class}} object.
}
\description{
Convert from \code{\link[S4Vectors]{SimpleList-class}}
to the multi-experiment data container

Generic Accessor Functions
}

