% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elist-class.R
\name{Elist}
\alias{Elist}
\title{Elist Acessor function for the \code{Elist} slot of a 
\code{MultiAssayExperiment} object}
\usage{
Elist(x)
}
\arguments{
\item{x}{A code{MultiAssayExperiment} class object}
}
\value{
A \code{Elist} class object of experiment data
}
\description{
Elist Acessor function for the \code{Elist} slot of a 
\code{MultiAssayExperiment} object
}

