% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\docType{methods}
\name{elist}
\alias{elist}
\alias{elist,MultiAssayExperiment-method}
\title{Generic Accessor Functions}
\usage{
elist(x)

\S4method{elist}{MultiAssayExperiment}(x)
}
\arguments{
\item{x}{A \code{\linkS4class{MultiAssayExperiment}} object.}
}
\value{
A \code{\link[S4Vectors]{SimpleList-class}} object.
}
\description{
Generic Accessor Functions
}
\section{Methods (by class)}{
\itemize{
\item \code{MultiAssayExperiment}: 
}}

