% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elist-class.R
\name{Elist}
\alias{Elist}
\title{Elist Acessor function for the \code{Elist} slot of a 
\code{MultiAssayExperiment} object}
\usage{
Elist(x)
}
\arguments{
\item{x}{A code{MultiAssayExperiment} class object}
}
\value{
A \code{Elist} class object of experiment data
}
\description{
Elist Acessor function for the \code{Elist} slot of a 
\code{MultiAssayExperiment} object
}
\examples{
## Create an empty Elist instance 
Elist()

## Create array matrix and AnnotatedDataFrame to create an ExpressionSet class
arraydat <- matrix(seq(101, 108), ncol=4,
                   dimnames = list(
                     c("ENST00000294241", "ENST00000355076"), 
                     c("array1", "array2", "array3", "array4")
                   ))
arraypdat <- as(data.frame(
  slope53 = rnorm(4), 
  row.names = c("array1", "array2", "array3", "array4")), 
  "AnnotatedDataFrame")
exprdat <- Biobase::ExpressionSet(assayData=arraydat, phenoData=arraypdat)

## Create a sample methylation dataset
methyldat <- matrix(1:10, ncol = 5,
                    dimnames = list(
                      c("ENST00000355076", "ENST00000383706"),
                      c("methyl1", "methyl2", "methyl3", "methyl4", "methyl5")))

## Combine to a named list and call the Elist constructor function
ExpList <- list(exprdat, methyldat)
names(ExpList) <- c("Affy", "Methyl450k")
myElist <- Elist(ExpList)
}

