% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\name{elist}
\alias{elist}
\title{Generic Accessor Functions}
\usage{
elist(x)
}
\arguments{
\item{x}{A \code{\linkS4class{MultiAssayExperiment}} object.}
}
\value{
A \code{\linkS4class{"SimpleList"}} object.
}
\description{
Generic Accessor Functions
}

