% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentList-class.R,
%   R/MultiAssayExperiment-methods.R, R/assay-methods.R
\docType{class}
\name{ExperimentList-class}
\alias{ExperimentList-class}
\alias{.ExperimentList}
\alias{ExperimentList,ANY-method}
\alias{ExperimentList,missing-method}
\alias{show,ExperimentList-method}
\alias{dimnames,ExperimentList-method}
\alias{reduce,ExperimentList-method}
\alias{assay,ANY,missing-method}
\alias{assays,ExperimentList-method}
\alias{assay,ExperimentList,missing-method}
\title{A container for multi-experiment data}
\usage{
\S4method{ExperimentList}{ANY}(x)

\S4method{ExperimentList}{missing}(x)

\S4method{show}{ExperimentList}(object)

\S4method{dimnames}{ExperimentList}(x)

\S4method{reduce}{ExperimentList}(x, drop.empty.ranges = FALSE, simplify,
  replicates = NULL, ...)

\S4method{assay}{ANY,missing}(x, i, ...)

\S4method{assays}{ExperimentList}(x, ..., withDimnames = TRUE)
}
\arguments{
\item{x}{constructor: A \code{list} object, reduce or assay: an
\code{ExperimentList} object}

\item{object}{An \code{\linkS4class{ExperimentList}} object}

\item{drop.empty.ranges}{unused argument}

\item{simplify}{A function for merging columns where duplicates are indicated
by replicates}

\item{replicates}{reduce: A \code{list} or \linkS4class{LogicalList} where
each element represents a sample and a vector of repeated measurements for
the sample (default NULL)}

\item{...}{Additional arguments. See details for more information.}

\item{withDimnames}{logical (default TRUE) whether to return dimension names}
}
\value{
An \code{ExperimentList} object
}
\description{
The \code{ExperimentList} class is a container that builds on
the \code{SimpleList} with additional
checks for consistency in experiment names and length.
It contains a \code{SimpleList} of experiments with sample identifiers.
One element present per experiment performed.
}
\details{
Convert from \code{SimpleList} or \code{list}
to the multi-experiment data container. When using the \strong{reduce}
method, additional arguments are passed to the given \link{combine}
function argument (e.g., na.rm = TRUE)
}
\section{Methods (by generic)}{
\itemize{
\item \code{ExperimentList}: Create an \code{ExperimentList} object from an
"ANY" class object, mainly \code{list}

\item \code{ExperimentList}: Create an empty ExperimentList for signature
"missing"

\item \code{show}: Show method for
\code{\linkS4class{ExperimentList}} class

\item \code{dimnames}: Get the dimension names for
an \code{ExperimentList} using \code{\link[IRanges]{CharacterList}}

\item \code{reduce}: Apply the reduce method on the
ExperimentList elements

\item \code{assay}: Get the assay data for the default ANY class

\item \code{assays}: Get the assay data from each element in the
\link{ExperimentList}
}}

\examples{
ExperimentList()

}
