% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentList-class.R,
%   R/MultiAssayExperiment-methods.R, R/assay-methods.R
\docType{class}
\name{ExperimentList-class}
\alias{ExperimentList-class}
\alias{.ExperimentList}
\alias{ExperimentList,ANY-method}
\alias{ExperimentList,missing-method}
\alias{show,ExperimentList-method}
\alias{dimnames,ExperimentList-method}
\alias{reduce,ExperimentList-method}
\alias{assay,ANY,missing-method}
\alias{assay,ExperimentList,missing-method}
\title{A container for multi-experiment data}
\usage{
\S4method{ExperimentList}{ANY}(x)

\S4method{ExperimentList}{missing}(x)

\S4method{show}{ExperimentList}(object)

\S4method{dimnames}{ExperimentList}(x)

\S4method{reduce}{ExperimentList}(x, drop.empty.ranges = FALSE,
  replicates = NULL, combine = rowMeans, vectorized = TRUE, ...)

\S4method{assay}{ANY,missing}(x, i)

\S4method{assay}{ExperimentList,missing}(x, i)
}
\arguments{
\item{x}{A \code{list} object}

\item{object}{An \code{\linkS4class{ExperimentList}} class object}

\item{drop.empty.ranges}{Ignored until further notice}

\item{replicates}{reduce: A \code{list} or \linkS4class{LogicalList} where
each element represents a sample and a vector of repeated experiments for
the sample (default NULL)}

\item{combine}{reduce: A function for consolidating columns in the matrix
representation of the data}

\item{vectorized}{reduce: (default TRUE) whether the \code{combine} function
is vectorized, optimized for working down the vector pairs}

\item{...}{Additional arguments. See details for more information.}

\item{i}{missing argument}
}
\value{
An \code{ExperimentList} class object
}
\description{
The \code{ExperimentList} class is a container that builds on
the \code{SimpleList} with additional
checks for consistency in experiment names and length.
It contains a \code{SimpleList} of experiments with sample identifiers.
One element present per experiment performed.
}
\details{
Convert from \code{SimpleList} or \code{list}
to the multi-experiment data container. When using the reduce method,
additional arguments are passed to the given \link{combine} function
argument (e.g., na.rm = TRUE)
}
\section{Methods (by generic)}{
\itemize{
\item \code{ExperimentList}: Create an \code{ExperimentList} object from an
"ANY" class object, mainly \code{list}

\item \code{ExperimentList}: Create an empty ExperimentList for signature
"missing"

\item \code{show}: Show method for
\code{\linkS4class{ExperimentList}} class

\item \code{dimnames}: Get the dimension names for
a \code{MultiAssayExperiment} using
\code{\link[IRanges]{CharacterList}}

\item \code{reduce}: Apply the reduce method on the
ExperimentList elements

\item \code{assay}: Get the assay data for the default ANY class

\item \code{assay}: Get the assay data from each element in the
\link{ExperimentList}
}}
\examples{
ExperimentList()

}

