% Generated by roxygen2: do not edit by hand
% Please edit documentation in:
% R/ExperimentList-class.R,
% R/MultiAssayExperiment-methods.R,
% R/assay-methods.R
\docType{class}
\name{ExperimentList-class}
\alias{.ExperimentList}
\alias{ExperimentList,ANY-method}
\alias{ExperimentList,missing-method}
\alias{ExperimentList-class}
\alias{assay,ANY,ANY-method}
\alias{assay,ExperimentList,ANY-method}
\alias{colnames,ExperimentList-method}
\alias{rownames,ExperimentList-method}
\alias{show,ExperimentList-method}
\title{A container for multi-experiment data}
\usage{
\S4method{ExperimentList}{ANY}(x)

\S4method{ExperimentList}{missing}(x)

\S4method{show}{ExperimentList}(object)

\S4method{rownames}{ExperimentList}(x)

\S4method{colnames}{ExperimentList}(x)

\S4method{assay}{ANY,ANY}(x)

\S4method{assay}{ExperimentList,ANY}(x)
}
\arguments{
\item{x}{A \code{list} object}

\item{object}{An \code{\linkS4class{ExperimentList}} class object}
}
\value{
An \code{ExperimentList} class object
}
\description{
The \code{ExperimentList} class is a container that builds on
the \code{SimpleList} with additional
checks for consistency in experiment names and length.
It contains a \code{SimpleList} of experiments with sample identifiers.
One element present per experiment performed.
}
\details{
Convert from \code{SimpleList} or \code{list}
to the multi-experiment data container
}
\section{Methods (by generic)}{
\itemize{
\item \code{ExperimentList}: Create an \code{ExperimentList} object from an
"ANY" class object, mainly \code{list}

\item \code{ExperimentList}: Create an empty ExperimentList for signature
"missing"

\item \code{show}: Show method for
\code{\linkS4class{ExperimentList}} class

\item \code{rownames}: Get all the rownames of an \code{ExperimentList}

\item \code{colnames}: Get sample names from an \code{ExperimentList}
object

\item \code{assay}: Get the assay data for the default ANY class

\item \code{assay}: Get the assay data from each element in the
\link{ExperimentList}
}}
\examples{
ExperimentList()

}

