% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biocMultiAssay-methods.R
\docType{methods}
\name{featExtractor}
\alias{featExtractor}
\alias{featExtractor,ExpressionSet-method}
\alias{featExtractor,GRangesList-method}
\alias{featExtractor,SummarizedExperiment-method}
\alias{featExtractor,matrix-method}
\title{Feature extractor method}
\usage{
featExtractor(x)

\S4method{featExtractor}{ExpressionSet}(x)

\S4method{featExtractor}{SummarizedExperiment}(x)

\S4method{featExtractor}{matrix}(x)

\S4method{featExtractor}{GRangesList}(x)
}
\arguments{
\item{x}{Either an \code{\linkS4class{ExpressionSet}}, \code{\linkS4class{GRangesList}}, \code{\linkS4class{SummarizedExperiment}} or \code{matrix} class object}
}
\value{
Returns either rownames or featureNames
}
\description{
Feature extractor method
}
\section{Methods (by class)}{
\itemize{
\item \code{ExpressionSet}: 

\item \code{SummarizedExperiment}: 

\item \code{matrix}: 

\item \code{GRangesList}: 
}}

