% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biocMultiAssay-methods.R
\docType{methods}
\name{features}
\alias{features}
\alias{features,ExpressionSet-method}
\alias{features,GRangesList-method}
\alias{features,MultiAssayExperiment-method}
\alias{features,RangedSummarizedExperiment-method}
\alias{features,matrix-method}
\title{Feature extractor method}
\usage{
features(x)

\S4method{features}{ExpressionSet}(x)

\S4method{features}{RangedSummarizedExperiment}(x)

\S4method{features}{matrix}(x)

\S4method{features}{GRangesList}(x)

\S4method{features}{MultiAssayExperiment}(x)
}
\arguments{
\item{x}{Either an \code{\linkS4class{ExpressionSet}}, \code{\linkS4class{GRangesList}}, \code{\linkS4class{RangedSummarizedExperiment}} or \code{matrix} class object}
}
\value{
Returns either rownames or featureNames
}
\description{
Feature extractor method
}
\section{Methods (by class)}{
\itemize{
\item \code{ExpressionSet}: 

\item \code{RangedSummarizedExperiment}: 

\item \code{matrix}: 

\item \code{GRangesList}: 

\item \code{MultiAssayExperiment}: 
}}

