% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-methods.R
\docType{methods}
\name{getHits}
\alias{getHits}
\alias{getHits,ANY,GRanges-method}
\alias{getHits,ANY,character-method}
\alias{getHits,GRanges,GRanges-method}
\alias{getHits,MultiAssayExperiment,GRanges-method}
\alias{getHits,MultiAssayExperiment,character-method}
\alias{getHits,RangedSummarizedExperiment,GRanges-method}
\title{Find hits by class type}
\usage{
getHits(subject, query, ...)

\S4method{getHits}{MultiAssayExperiment,character}(subject, query, ...)

\S4method{getHits}{MultiAssayExperiment,GRanges}(subject, query, ...)

\S4method{getHits}{GRanges,GRanges}(subject, query, ...)

\S4method{getHits}{ANY,GRanges}(subject, query, ...)

\S4method{getHits}{RangedSummarizedExperiment,GRanges}(subject, query, ...)

\S4method{getHits}{ANY,character}(subject, query, ...)
}
\arguments{
\item{subject}{Any valid element from the \code{\linkS4class{Elist}} class}

\item{query}{Either a \code{character} vector or
\code{\linkS4class{GRanges}}
object used to search by name or ranges}

\item{...}{Additional arguments to findOverlaps}
}
\value{
Names of matched queries
}
\description{
Find hits by class type
}
\section{Methods (by class)}{
\itemize{
\item \code{subject = MultiAssayExperiment,query = character}: Find all matching rownames by \code{character}

\item \code{subject = MultiAssayExperiment,query = GRanges}: Find all matching rownames by \code{GRanges}

\item \code{subject = GRanges,query = GRanges}: Find and get corresponding names of two \code{GRanges}
using \code{findOverlaps}

\item \code{subject = ANY,query = GRanges}: Find all matching rownames for range-based objects

\item \code{subject = RangedSummarizedExperiment,query = GRanges}: Find rownames for \code{RangedSummarizedExperiment} hits

\item \code{subject = ANY,query = character}: Find all matching rownames based on \code{character} query
}}
\examples{
## Load an example MultiAssayExperiment object
example("MultiAssayExperiment")
example("GRangesList")

## Find what ranges fit the criteria (see findOverlaps)
getHits(myMultiAssayExperiment, gr1)
}

