% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocMultiAssay-methods.R
\docType{methods}
\name{getMap,Stage-method}
\alias{getMap,Stage-method}
\title{describeIn Stage Convert map from list to DataFrame}
\usage{
\S4method{getMap}{Stage}(object)
}
\description{
describeIn Stage Convert map from list to DataFrame
}

