% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocMultiAssay-methods.R
\name{getMap}
\alias{getMap}
\title{Convert MultiAssayView slot "keeps" to Map}
\usage{
getMap(object)
}
\arguments{
\item{object}{An \linkS4class{MultiAssayView} class object}
}
\value{
Returns a data.frame representation of samples
}
\description{
Convert MultiAssayView slot "keeps" to Map
}

