% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Identify-class.R
\docType{class}
\name{Identify-class}
\alias{Identify-class}
\title{An identifier class used for subsetting}
\description{
An identifier class used for subsetting
}
\section{Slots}{

\describe{
\item{\code{logreturn}}{A \code{logical} vector for subsetting}

\item{\code{drops}}{A \code{list} of excluded information due to subsetting}
}}

