% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify-class.R
\docType{class}
\name{Identify-class}
\alias{Identify-class}
\title{An identifier class used for subsetting}
\description{
An identifier class used for subsetting
}
\section{Slots}{

\describe{
\item{\code{query}}{Any class indicator needed to subset}

\item{\code{keeps}}{A \code{list} indicating valid matches in each assay}

\item{\code{drops}}{A \code{list} of excluded information due to subsetting}

\item{\code{type}}{A \code{character} vector indicating method used to search}
}}

