% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\name{masterPheno}
\alias{masterPheno}
\title{Generic Accessor Functions}
\usage{
masterPheno(x)
}
\arguments{
\item{x}{A \code{\linkS4class{MultiAssayExperiment}} object.}
}
\value{
A \code{"data.frame"} object.
}
\description{
Generic Accessor Functions
}

