% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\docType{methods}
\name{masterPheno}
\alias{masterPheno}
\alias{masterPheno,MultiAssayExperiment-method}
\title{Generic Accessor Functions}
\usage{
masterPheno(x)

\S4method{masterPheno}{MultiAssayExperiment}(x)
}
\arguments{
\item{x}{A \code{\link{MultiAssayExperiment}} object}
}
\value{
A \code{data.frame} object
}
\description{
Generic Accessor Functions
}
\section{Methods (by class)}{
\itemize{
\item \code{MultiAssayExperiment}: 
}}

