% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\docType{class}
\name{MatchedAssayExperiment-class}
\alias{MatchedAssayExperiment-class}
\alias{MatchedAssayExperiment}
\title{An integrative and matched-samples class for experiment data}
\usage{
MatchedAssayExperiment(experiments = ExperimentList(),
  colData = S4Vectors::DataFrame(), sampleMap = S4Vectors::DataFrame(assay =
  factor(), primary = character(), colname = character()), metadata = NULL,
  drops = list())
}
\arguments{
\item{experiments}{A \code{list} or \link{ExperimentList} of all
combined experiments}

\item{colData}{A \code{\link[S4Vectors]{DataFrame}} or \code{data.frame} of
characteristics for all biological units}

\item{sampleMap}{A \code{DataFrame} or \code{data.frame} of assay names,
sample identifiers, and colname samples}

\item{metadata}{An optional argument of "ANY" class (usually list) for
content describing the experiments}

\item{drops}{A \code{list} of unmatched information
(included after subsetting)}
}
\value{
A \code{MatchedAssayExperiment} object
}
\description{
This class supports the use of matched samples where an equal number
of observations per biological unit are present in all assays.
}
\section{Functions}{
\itemize{
\item \code{MatchedAssayExperiment}: Construct a
\code{MatchedAssayExperiment} class from \linkS4class{MultiAssayExperiment}
inputs.
}}

\examples{
MatchedAssayExperiment()

}
\seealso{
MultiAssayExperiment
}
