% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\docType{class}
\name{MatchedAssayExperiment-class}
\alias{MatchedAssayExperiment-class}
\alias{MatchedAssayExperiment}
\alias{coerce,MultiAssayExperiment,MatchedAssayExperiment-method}
\title{An integrative and matched-samples class for experiment data}
\usage{
MatchedAssayExperiment(...)
}
\arguments{
\item{...}{Either a single MultiAssayExperiment or the components to create
a valid MultiAssayExperiment}
}
\value{
A \code{MatchedAssayExperiment} object
}
\description{
This class supports the use of matched samples where an equal number
of observations per biological unit are present in all assays.
}
\section{Functions}{
\itemize{
\item \code{MatchedAssayExperiment}: Construct a
\code{MatchedAssayExperiment} class from \linkS4class{MultiAssayExperiment}
inputs.
}}

\examples{
data("miniACC")
acc <- as(miniACC, "MatchedAssayExperiment")
acc

}
\seealso{
\link{MultiAssayExperiment}
}
