% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R, R/subset.R
\docType{class}
\name{MultiAssayExperiment-class}
\alias{MultiAssayExperiment-class}
\alias{elist,MultiAssayExperiment-method}
\alias{length,MultiAssayExperiment-method}
\alias{masterPheno,MultiAssayExperiment-method}
\alias{metadata,MultiAssayExperiment-method}
\alias{names,MultiAssayExperiment-method}
\alias{sampleMap,MultiAssayExperiment-method}
\alias{subset,MultiAssayExperiment-method}
\title{An integrative MultiAssay class for experiment data}
\usage{
\S4method{sampleMap}{MultiAssayExperiment}(x)

\S4method{elist}{MultiAssayExperiment}(x)

\S4method{masterPheno}{MultiAssayExperiment}(x)

\S4method{metadata}{MultiAssayExperiment}(x)

\S4method{length}{MultiAssayExperiment}(x)

\S4method{names}{MultiAssayExperiment}(x)

\S4method{subset}{MultiAssayExperiment}(x, indicator, by, ...)
}
\arguments{
\item{x}{A \code{\link{MultiAssayExperiment}} object.}

\item{indicator}{A \code{logical} or \code{character} vector or \code{Identify} class object to use for subsetting}

\item{by}{A \code{character} vector of length one designating to subset either by samples, features, or assays}

\item{x}{A \code{\link{MultiAssayExperiment}} object}

\item{x}{A \code{\link{MultiAssayExperiment}} object}

\item{x}{A \code{\link{MultiAssayExperiment}} object}

\item{MultiAssay}{A \code{\linkS4class{MultiAssayExperiment}} object}
}
\value{
A \code{\link[S4Vectors]{SimpleList-class}} object.

Any type of object describing the metadata

An \code{integer}

A character vector of experiment names
}
\description{
An integrative MultiAssay class for experiment data

Generic Accessor Functions

Generic Acessor Functions

Length of Experiment List

Names of Experiments

Subset MultiAssayExperiment object
\code{subset} returns a subsetted \code{\linkS4class{MultiAssayExperiment}} object
}
\section{Methods (by generic)}{
\itemize{
\item \code{sampleMap}: Access sampleMap slot from MultiAssayExperiment

\item \code{elist}: Access elist class from MultiAssayExperiment

\item \code{masterPheno}: Access masterPheno slot from MultiAssayExperiment

\item \code{metadata}: Access metadata slot from MultiAssayExperiment

\item \code{length}: Get the length of elist

\item \code{names}: Get the names of the elist

\item \code{subset}: export subset
}}
\section{Slots}{

\describe{
\item{\code{elist}}{A \code{\linkS4class{elist}} class object for each assay dataset.}

\item{\code{masterPheno}}{A \code{data.frame} of all clinical data available across experiments.}

\item{\code{sampleMap}}{A \code{data.frame} of translatable identifiers of samples and participants.}

\item{\code{metadata}}{Additional data describing the \code{\link{MultiAssayExperiment}} object.}

\item{\code{drops}}{A metadata \code{list} of dropped information.}
}}

