% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\docType{class}
\name{MultiAssayExperiment-class}
\alias{MultiAssayExperiment-class}
\alias{length,MultiAssayExperiment-method}
\alias{names,MultiAssayExperiment-method}
\title{An integrative MultiAssay class for experiment data}
\usage{
\S4method{length}{MultiAssayExperiment}(x)

\S4method{names}{MultiAssayExperiment}(x)
}
\arguments{
\item{x}{A \code{\link{MultiAssayExperiment}} object}

\item{x}{A \code{\link{MultiAssayExperiment}} object}
}
\value{
An \code{integer}

A character vector of experiment names
}
\description{
An integrative MultiAssay class for experiment data

Length of Experiment List

Names of Experiments
}
\section{Methods (by generic)}{
\itemize{
\item \code{length}: 

\item \code{names}: 
}}
\section{Slots}{

\describe{
\item{\code{elist}}{A \code{\link[S4Vectors]{SimpleList-class}} of data across different types of assays.}

\item{\code{masterPheno}}{A \code{data.frame} of all clinical data available across experiments.}

\item{\code{sampleMap}}{A \code{data.frame} of translatable identifiers of samples and participants.}

\item{\code{metadata}}{Additional data describing the \code{\link{MultiAssayExperiment}} object.}
}}

