% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R, R/MultiAssayExperiment-methods.R
\docType{class}
\name{MultiAssayExperiment-class}
\alias{Elist,MultiAssayExperiment-method}
\alias{Elist<-,MultiAssayExperiment,Elist-method}
\alias{MultiAssayExperiment-class}
\alias{[,MultiAssayExperiment,ANY-method}
\alias{assay,MultiAssayExperiment,ANY-method}
\alias{colnames,MultiAssayExperiment-method}
\alias{isEmpty,MultiAssayExperiment-method}
\alias{length,MultiAssayExperiment-method}
\alias{metadata,MultiAssayExperiment-method}
\alias{names,MultiAssayExperiment-method}
\alias{pData,MultiAssayExperiment-method}
\alias{rownames,MultiAssayExperiment-method}
\alias{sampleMap,MultiAssayExperiment-method}
\alias{sampleMap<-,MultiAssayExperiment,DataFrame-method}
\alias{show,MultiAssayExperiment-method}
\title{An integrative MultiAssay class for experiment data}
\usage{
\S4method{show}{MultiAssayExperiment}(object)

\S4method{sampleMap}{MultiAssayExperiment}(x)

\S4method{Elist}{MultiAssayExperiment}(x)

\S4method{pData}{MultiAssayExperiment}(object)

\S4method{metadata}{MultiAssayExperiment}(x)

\S4method{length}{MultiAssayExperiment}(x)

\S4method{names}{MultiAssayExperiment}(x)

\S4method{sampleMap}{MultiAssayExperiment,DataFrame}(x) <- value

\S4method{Elist}{MultiAssayExperiment,Elist}(x) <- value

\S4method{rownames}{MultiAssayExperiment}(x)

\S4method{colnames}{MultiAssayExperiment}(x)

\S4method{assay}{MultiAssayExperiment,ANY}(x)

\S4method{[}{MultiAssayExperiment,ANY}(x, i, j, k, ..., drop = TRUE)

\S4method{isEmpty}{MultiAssayExperiment}(x)
}
\arguments{
\item{object}{A \code{MultiAssayExperiment} class object}

\item{x}{A \code{MultiAssayExperiment} object for subsetting}

\item{value}{A \code{DataFrame} or \code{Elist} object to replace the existing
\code{sampleMap} or an \code{Elist} slot, respectively}

\item{i}{Either a \code{character}, or \code{GRanges} object for subsetting
by rows}

\item{j}{Either a \code{character}, \code{logical}, or \code{numeric} vector
for subsetting by columns}

\item{k}{Either a \code{character}, \code{logical}, or \code{numeric} vector
for subsetting by assays}

\item{...}{Additional arguments passed down to \code{getHits} support
function for subsetting by rows}

\item{drop}{logical (default TRUE) whether to drop empty assay elements
in the \code{Elist}}
}
\value{
A \code{MultiAssayExperiment} object
}
\description{
The \code{MultiAssayExperiment} class can be used to manage results of 
diverse assays on a collection of specimen. Currently,  the class can handle
assays that are organized instances of \code{RangedSummarizedExperiment}, 
\code{ExpressionSet}, \code{matrix}, \code{RangedRaggedAssay} 
(inherits from \code{GRangesList}), and \code{RangedVcfStack}. Create new
\code{MultiAssayExperiment} instances with the eponymous constructor, 
minimally with the argument \code{\linkS4class{Elist}}, potentially also
with the arguments \code{pData} and \code{sampleMap}
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Show method for a
\code{MultiAssayExperiment}

\item \code{sampleMap}: Access sampleMap slot from
MultiAssayExperiment

\item \code{Elist}: Access Elist class from
MultiAssayExperiment

\item \code{pData}: Access pData slot from
MultiAssayExperiment

\item \code{metadata}: Access metadata slot from
MultiAssayExperiment

\item \code{length}: Get the length of Elist

\item \code{names}: Get the names of the Elist

\item \code{sampleMap<-}: value: A \code{DataFrame} sampleMap 
representation

\item \code{Elist<-}: value: An \code{Elist} 
representation

\item \code{rownames}: Get all the rownames for a
MultiAssayExperiment using \code{\link[IRanges]{CharacterList}}

\item \code{colnames}: Get all the colnames for a
MultiAssayExperiment

\item \code{assay}: Get the raw data from a
MultiAssayExperiment as a list

\item \code{[}: Subset a \code{MultiAssayExperiment} object

\item \code{isEmpty}: Logical value of empty
\code{MultiAssayExperiment}
}}
\section{Slots}{

\describe{
\item{\code{Elist}}{A \code{\linkS4class{Elist}} class object for each assay dataset}

\item{\code{pData}}{A \code{DataFrame} of all clinical data available across
experiments}

\item{\code{sampleMap}}{A \code{DataFrame} of translatable identifiers of samples
and participants}

\item{\code{metadata}}{Additional data describing the
\code{\link{MultiAssayExperiment}} object}

\item{\code{drops}}{A metadata \code{list} of dropped information.}
}}
\examples{
## Run the example Elist
example("Elist-class")

## Run the sample map example
example("listToMap")

## Create an example phenotype data
myPheno <- data.frame(sex = c("M", "F", "M", "F"),
                       age = 38:41,
                       row.names = c("Jack", "Jill", "Bob", "Barbara"))

## Auto-creation of map
myMultiAssayExperiment <- MultiAssayExperiment(Elist = myElist,
                                               pData = myPheno,
                                               sampleMap = myDF)
}
\seealso{
\code{getHits}
}

