% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\docType{class}
\name{MultiAssayExperiment-class}
\alias{MultiAssayExperiment-class}
\title{An integrative MultiAssay class for experiment data}
\description{
An integrative MultiAssay class for experiment data
}
\section{Slots}{

\describe{
\item{\code{elist}}{A list of data across different types of assays}

\item{\code{masterPheno}}{A data.frame of all clinical data available across experiments}

\item{\code{sampleMap}}{A list of translatable identifiers of samples and participants}

\item{\code{metadata}}{Additional data describing the \code{\linkS4class{MultiAssayExperiment}} class}
}}

