% Generated by roxygen2: do not edit by hand
% Please edit documentation in:
% R/MultiAssayExperiment-class.R,
% R/MultiAssayExperiment-methods.R,
% R/assay-methods.R
\docType{class}
\name{MultiAssayExperiment-class}
\alias{$,MultiAssayExperiment-method}
\alias{$<-,MultiAssayExperiment-method}
\alias{MultiAssayExperiment-class}
\alias{[,MultiAssayExperiment,ANY,ANY,ANY-method}
\alias{[,MultiAssayExperiment,ANY-method}
\alias{assay,MultiAssayExperiment,missing-method}
\alias{colnames,MultiAssayExperiment-method}
\alias{complete.cases,MultiAssayExperiment-method}
\alias{experiments,MultiAssayExperiment-method}
\alias{experiments<-,MultiAssayExperiment,ExperimentList-method}
\alias{isEmpty,MultiAssayExperiment-method}
\alias{length,MultiAssayExperiment-method}
\alias{metadata,MultiAssayExperiment-method}
\alias{metadata<-,MultiAssayExperiment-method}
\alias{names,MultiAssayExperiment-method}
\alias{pData,MultiAssayExperiment-method}
\alias{pData<-,MultiAssayExperiment,DataFrame-method}
\alias{rownames,MultiAssayExperiment-method}
\alias{sampleMap,MultiAssayExperiment-method}
\alias{sampleMap<-,MultiAssayExperiment,DataFrame-method}
\alias{show,MultiAssayExperiment-method}
\alias{updateObject,MultiAssayExperiment-method}
\title{An integrative MultiAssay class for experiment data}
\usage{
\S4method{show}{MultiAssayExperiment}(object)

\S4method{sampleMap}{MultiAssayExperiment}(x)

\S4method{experiments}{MultiAssayExperiment}(x)

\S4method{pData}{MultiAssayExperiment}(object)

\S4method{metadata}{MultiAssayExperiment}(x)

\S4method{length}{MultiAssayExperiment}(x)

\S4method{names}{MultiAssayExperiment}(x)

\S4method{sampleMap}{MultiAssayExperiment,DataFrame}(object) <- value

\S4method{experiments}{MultiAssayExperiment,ExperimentList}(object) <- value

\S4method{pData}{MultiAssayExperiment,DataFrame}(object) <- value

\S4method{metadata}{MultiAssayExperiment}(x, ...) <- value

\S4method{$}{MultiAssayExperiment}(x, name) <- value

\S4method{updateObject}{MultiAssayExperiment}(object, ..., verbose = FALSE)

\S4method{rownames}{MultiAssayExperiment}(x)

\S4method{colnames}{MultiAssayExperiment}(x)

\S4method{$}{MultiAssayExperiment}(x, name)

\S4method{[}{MultiAssayExperiment,ANY,ANY,ANY}(x, i, j, k, ..., drop = TRUE)

\S4method{isEmpty}{MultiAssayExperiment}(x)

\S4method{complete.cases}{MultiAssayExperiment}(...)

\S4method{assay}{MultiAssayExperiment,missing}(x, i)
}
\arguments{
\item{object}{A \code{MultiAssayExperiment} class object}

\item{x}{A \code{MultiAssayExperiment} object for subsetting}

\item{value}{A \code{DataFrame} or \code{ExperimentList} object to replace
the existing \code{sampleMap}, \code{ExperimentList}, or \code{pData} slot}

\item{...}{Additional arguments passed down to \code{getHits} support
function for subsetting by rows}

\item{name}{pData column name}

\item{verbose}{(logical default FALSE) whether to output verbose}

\item{i}{Either a \code{character}, or \code{GRanges} object for subsetting
by rows}

\item{j}{Either a \code{character}, \code{logical}, or \code{numeric} vector
for subsetting by columns}

\item{k}{Either a \code{character}, \code{logical}, or \code{numeric} vector
for subsetting by assays}

\item{drop}{logical (default TRUE) whether to drop empty assay elements
in the \code{ExperimentList}}
}
\value{
A \code{MultiAssayExperiment} object
}
\description{
The \code{MultiAssayExperiment} class can be used to manage results of
diverse assays on a collection of specimen. Currently,  the class can handle
assays that are organized instances of
\code{\linkS4class{SummarizedExperiment}},
\code{\linkS4class{ExpressionSet}},
\code{matrix}, \code{\link{RangedRaggedAssay}}
(inherits from \code{\linkS4class{GRangesList}}), and \code{RangedVcfStack}.
Create new \code{MultiAssayExperiment} instances with the eponymous
constructor, minimally with the argument \code{\link{ExperimentList}},
potentially also with the arguments \code{pData} (see section below) and
\code{\link{sampleMap}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Show method for a
\code{MultiAssayExperiment}

\item \code{sampleMap}: Access sampleMap slot from a
MultiAssayExperiment

\item \code{experiments}: Access ExperimentList class from a
MultiAssayExperiment

\item \code{pData}: Access pData slot from a
MultiAssayExperiment

\item \code{metadata}: Access metadata slot from a
MultiAssayExperiment

\item \code{length}: Get the length of ExperimentList

\item \code{names}: Get the names of the ExperimentList

\item \code{sampleMap<-}: value: A \code{DataFrame} sampleMap
representation

\item \code{experiments<-}: value: An \code{ExperimentList}
representation

\item \code{pData<-}: value: A \code{DataFrame} of specimen data

\item \code{metadata<-}: value: Data of type "ANY"

\item \code{$<-}: value: DataFrame column

\item \code{updateObject}: Update old serialized MultiAssayExperiment
objects to new API

\item \code{rownames}: Get all the rownames for a
\code{MultiAssayExperiment} using a \code{\link[IRanges]{CharacterList}}

\item \code{colnames}: Get all the colnames for a
\code{MultiAssayExperiment}

\item \code{$}: Access pData column

\item \code{[}: Subset a \code{MultiAssayExperiment} object

\item \code{isEmpty}: A \code{logical} value indicating an empty
\code{MultiAssayExperiment}

\item \code{complete.cases}: Return a logical vector of biological units
with data across all experiments

\item \code{assay}: Get the assay data for a
\link{MultiAssayExperiment} as a \code{list}
}}
\section{Slots}{

\describe{
\item{\code{ExperimentList}}{A \code{\link{ExperimentList}} class object for
each assay dataset}

\item{\code{pData}}{A \code{DataFrame} of all clinical/specimen data available
across experiments}

\item{\code{sampleMap}}{A \code{DataFrame} of translatable identifiers
of samples and participants}

\item{\code{metadata}}{Additional data describing the
\code{MultiAssayExperiment} object}

\item{\code{drops}}{A metadata \code{list} of dropped information}
}}
\section{pData}{

The \code{pData} slot is a collection of primary specimen data valid across
all experiments. This slot is strictly of class
\code{\linkS4class{DataFrame}} but arguments for the constructor function
allow arguments to be of class \code{data.frame} and subsequently coerced.
}

\section{ExperimentList}{

The \code{\link{ExperimentList}} slot is designed to contain results from
each experiment/assay. It contains a \link[S4Vectors]{SimpleList}.
}

\section{sampleMap}{

The \code{\link{sampleMap}} contains a \code{DataFrame} of translatable
identifiers of samples and participants or biological units. Standard column
names of the sampleMap are "assay", "primary", and "colname".
}
\examples{
MultiAssayExperiment()

}
\seealso{
\code{getHits}
}

