% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R,
%   R/MultiAssayExperiment-methods.R, R/assay-methods.R
\docType{class}
\name{MultiAssayExperiment-class}
\alias{MultiAssayExperiment-class}
\alias{show,MultiAssayExperiment-method}
\alias{sampleMap,MultiAssayExperiment-method}
\alias{experiments,MultiAssayExperiment-method}
\alias{pData,MultiAssayExperiment-method}
\alias{metadata,MultiAssayExperiment-method}
\alias{length,MultiAssayExperiment-method}
\alias{names,MultiAssayExperiment-method}
\alias{sampleMap<-,MultiAssayExperiment,DataFrame-method}
\alias{experiments<-,MultiAssayExperiment,ExperimentList-method}
\alias{pData<-,MultiAssayExperiment,DataFrame-method}
\alias{metadata<-,MultiAssayExperiment-method}
\alias{$<-,MultiAssayExperiment-method}
\alias{updateObject,MultiAssayExperiment-method}
\alias{dimnames,MultiAssayExperiment-method}
\alias{$,MultiAssayExperiment-method}
\alias{[,MultiAssayExperiment,ANY,ANY,ANY-method}
\alias{[,MultiAssayExperiment,ANY-method}
\alias{isEmpty,MultiAssayExperiment-method}
\alias{[[,MultiAssayExperiment,ANY,ANY-method}
\alias{[[<-,MultiAssayExperiment,ANY,ANY-method}
\alias{complete.cases,MultiAssayExperiment-method}
\alias{duplicated,MultiAssayExperiment-method}
\alias{reduce,MultiAssayExperiment-method}
\alias{reduce,ANY-method}
\alias{c,MultiAssayExperiment-method}
\alias{assay,MultiAssayExperiment,missing-method}
\title{An integrative multi-assay class for experiment data}
\usage{
\S4method{show}{MultiAssayExperiment}(object)

\S4method{sampleMap}{MultiAssayExperiment}(x)

\S4method{experiments}{MultiAssayExperiment}(x)

\S4method{pData}{MultiAssayExperiment}(object)

\S4method{metadata}{MultiAssayExperiment}(x)

\S4method{length}{MultiAssayExperiment}(x)

\S4method{names}{MultiAssayExperiment}(x)

\S4method{sampleMap}{MultiAssayExperiment,DataFrame}(object) <- value

\S4method{experiments}{MultiAssayExperiment,ExperimentList}(object) <- value

\S4method{pData}{MultiAssayExperiment,DataFrame}(object) <- value

\S4method{metadata}{MultiAssayExperiment}(x, ...) <- value

\S4method{$}{MultiAssayExperiment}(x, name) <- value

\S4method{updateObject}{MultiAssayExperiment}(object, ..., verbose = FALSE)

\S4method{dimnames}{MultiAssayExperiment}(x)

\S4method{$}{MultiAssayExperiment}(x, name)

\S4method{[}{MultiAssayExperiment,ANY,ANY,ANY}(x, i, j, k, ..., drop = TRUE)

\S4method{isEmpty}{MultiAssayExperiment}(x)

\S4method{[[}{MultiAssayExperiment,ANY,ANY}(x, i, j, ...)

\S4method{[[}{MultiAssayExperiment,ANY,ANY}(x, i, j, ...) <- value

\S4method{complete.cases}{MultiAssayExperiment}(...)

\S4method{duplicated}{MultiAssayExperiment}(x, incomparables = FALSE, ...)

\S4method{reduce}{MultiAssayExperiment}(x, drop.empty.ranges = FALSE,
  replicates = NULL, combine = rowMeans, vectorized = TRUE, ...)

\S4method{reduce}{ANY}(x, drop.empty.ranges = FALSE, replicates = NULL,
  combine = rowMeans, vectorized = TRUE, ...)

\S4method{c}{MultiAssayExperiment}(x, ..., sampleMap = NULL, mapFrom = NULL)

\S4method{assay}{MultiAssayExperiment,missing}(x, i, ...)
}
\arguments{
\item{object}{A \code{MultiAssayExperiment} class object}

\item{x}{A \code{MultiAssayExperiment} object for subsetting}

\item{value}{A \code{DataFrame} or \code{ExperimentList} object to replace
the existing \code{sampleMap}, \code{ExperimentList}, or \code{pData} slot}

\item{...}{Additional arguments. See details for more information.}

\item{name}{pData column name}

\item{verbose}{logical (default FALSE) whether to print extra messages}

\item{i}{subsetting: Either a \code{character}, or \code{GRanges} object for subsetting
by rows, assay: unused argument (missing)}

\item{j}{Either a \code{character}, \code{logical}, or \code{numeric} vector
for subsetting by columns}

\item{k}{Either a \code{character}, \code{logical}, or \code{numeric} vector
for subsetting by assays}

\item{drop}{logical (default TRUE) whether to drop empty assay elements
in the \code{ExperimentList}}

\item{incomparables}{duplicated: unused argument}

\item{drop.empty.ranges}{unused generic argument}

\item{replicates}{reduce: A list of \linkS4class{LogicalList} indicating
duplicate entries for each biological unit, see the \code{duplicated} method
for \code{MultiAssayExperiment}}

\item{combine}{reduce: function for combining replicate columns/samples
(default rowMeans)}

\item{vectorized}{reduce: logical (default TRUE) whether the combine function is
vectorized, optimized for working down the vector pairs}

\item{sampleMap}{\code{c} method: a \code{sampleMap} \code{list} or
\code{DataFrame} to guide merge}

\item{mapFrom}{Either a \code{logical}, \code{character}, or \code{integer}
vector indicating the experiment(s) that have an identical colname order as
the experiment input(s)}
}
\value{
A \code{MultiAssayExperiment} object

A \code{MultiAssayExperiment} object
}
\description{
The \code{MultiAssayExperiment} class can be used to manage results of
diverse assays on a collection of specimen. Currently,  the class can handle
assays that are organized instances of
\code{\linkS4class{SummarizedExperiment}},
\code{\linkS4class{ExpressionSet}},
\code{matrix}, \code{\link{RangedRaggedAssay}}
(inherits from \code{\linkS4class{GRangesList}}), and \code{RangedVcfStack}.
Create new \code{MultiAssayExperiment} instances with the eponymous
constructor, minimally with the argument \code{\link{ExperimentList}},
potentially also with the arguments \code{pData} (see section below) and
\code{\link{sampleMap}}.
}
\details{
The dots (\code{\ldots}) argument allows the user to specify additional
arguments in several instances. When subsetting (\strong{[}) a
\code{MultiAssayExperiment}, the dots allow for additional
arguments to be sent to \link{findOverlaps}. When using the \strong{reduce}
method, the dots are used to specify arguments for the supplied
\code{combine} argument and function. When using the \strong{assay} method,
additional arguments may be passed to the \code{RangedRaggedAssay} method.
See the link for more information:
\link{assay,RangedRaggedAssay,missing-method}. When using \strong{c} method
to add experiments to a \code{MultiAssayExperiment}, the dots allow extra
data classes compatible with the MultiAssayExperiment API. See: \link{API}
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Show method for a
\code{MultiAssayExperiment}

\item \code{sampleMap}: Access sampleMap slot from a
MultiAssayExperiment

\item \code{experiments}: Access ExperimentList class from a
MultiAssayExperiment

\item \code{pData}: Access pData slot from a
MultiAssayExperiment

\item \code{metadata}: Access metadata slot from a
MultiAssayExperiment

\item \code{length}: Get the length of ExperimentList

\item \code{names}: Get the names of the ExperimentList

\item \code{sampleMap<-}: value: A \code{DataFrame} sampleMap
representation

\item \code{experiments<-}: value: An \code{ExperimentList}
representation

\item \code{pData<-}: value: A \code{DataFrame} of specimen data

\item \code{metadata<-}: value: Data of type "ANY"

\item \code{$<-}: value: DataFrame column

\item \code{updateObject}: Update old serialized MultiAssayExperiment
objects to new API

\item \code{dimnames}: Get the dimension names
for a \code{MultiAssayExperiment} object

\item \code{$}: Access pData column

\item \code{[}: Subset a \code{MultiAssayExperiment} object

\item \code{isEmpty}: A \code{logical} value indicating an empty
\code{MultiAssayExperiment}

\item \code{[[}: Extract the \link{ExperimentList} element

\item \code{[[<-}: Replace the \link{ExperimentList} element
with a supported class (should have documented dimensions
in \code{sampleMap})

\item \code{complete.cases}: Return a logical vector of biological units
with data across all experiments

\item \code{duplicated}: Find duplicate columns in the data by
matching pData rownames

\item \code{reduce}: Housekeeping method for a
MultiAssayExperiment where only complete.cases are returned, replicate
measurements are averaged, and columns are aligned by the row order in pData.

\item \code{reduce}: Consolidate columns for rectangular
data structures, mainly matrix

\item \code{c}: Add an element to the
\code{ExperimentList} data slot

\item \code{assay}: Get the assay data for a
\link{MultiAssayExperiment} as a \code{list}
}}

\section{Slots}{

\describe{
\item{\code{ExperimentList}}{A \code{\link{ExperimentList}} class object for
each assay dataset}

\item{\code{pData}}{A \code{DataFrame} of all clinical/specimen data available
across experiments}

\item{\code{sampleMap}}{A \code{DataFrame} of translatable identifiers
of samples and participants}

\item{\code{metadata}}{Additional data describing the
\code{MultiAssayExperiment} object}

\item{\code{drops}}{A metadata \code{list} of dropped information}
}}

\section{pData}{

The \code{pData} slot is a collection of primary specimen data valid across
all experiments. This slot is strictly of class
\code{\linkS4class{DataFrame}} but arguments for the constructor function
allow arguments to be of class \code{data.frame} and subsequently coerced.
}

\section{ExperimentList}{

The \code{\link{ExperimentList}} slot is designed to contain results from
each experiment/assay. It contains a \link[S4Vectors]{SimpleList}.
}

\section{sampleMap}{

The \code{\link{sampleMap}} contains a \code{DataFrame} of translatable
identifiers of samples and participants or biological units. Standard column
names of the sampleMap are "assay", "primary", and "colname".
}

\examples{
example("MultiAssayExperiment")

## Subsetting
# Rows (i) Rows/Features in each experiment
myMultiAssayExperiment[1, , ]
myMultiAssayExperiment[c(TRUE, FALSE), , ]

# Columns (j) Rows in pData
myMultiAssayExperiment[, rownames(pData(myMultiAssayExperiment))[3:2],  ]

# Assays (k)
myMultiAssayExperiment[, , "Affy"]

## Complete cases (returns logical vector)
completes <- complete.cases(myMultiAssayExperiment)
compMAE <- myMultiAssayExperiment[, completes, ]
compMAE
pData(compMAE)

example("MultiAssayExperiment")

## Add an experiment
test1 <- myMultiAssayExperiment[[1L]]
colnames(test1) <- rownames(pData(myMultiAssayExperiment))

## Combine current MultiAssayExperiment with additional experiment
## (no sampleMap)
c(myMultiAssayExperiment, newExperiment = test1)

test2 <- myMultiAssayExperiment[[1L]]
c(myMultiAssayExperiment, newExp = test2, mapFrom = 3L)

}
