% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R, R/biocMultiAssay-methods.R, R/subset.R
\docType{class}
\name{MultiAssayExperiment-class}
\alias{Elist,MultiAssayExperiment-method}
\alias{MultiAssayExperiment-class}
\alias{[,RangedRaggedAssay,ANY-method}
\alias{[,RangedRaggedAssay,GRanges-method}
\alias{assay,MultiAssayExperiment,ANY-method}
\alias{colnames,MultiAssayExperiment-method}
\alias{length,MultiAssayExperiment-method}
\alias{masterPheno,MultiAssayExperiment-method}
\alias{metadata,MultiAssayExperiment-method}
\alias{names,MultiAssayExperiment-method}
\alias{rownames,MultiAssayExperiment-method}
\alias{sampleMap,MultiAssayExperiment-method}
\alias{subset,MultiAssayExperiment-method}
\title{An integrative MultiAssay class for experiment data}
\usage{
\S4method{sampleMap}{MultiAssayExperiment}(x)

\S4method{Elist}{MultiAssayExperiment}(x)

\S4method{masterPheno}{MultiAssayExperiment}(x)

\S4method{metadata}{MultiAssayExperiment}(x)

\S4method{length}{MultiAssayExperiment}(x)

\S4method{names}{MultiAssayExperiment}(x)

\S4method{rownames}{MultiAssayExperiment}(x)

\S4method{colnames}{MultiAssayExperiment}(x)

\S4method{assay}{MultiAssayExperiment,ANY}(x)

\S4method{[}{RangedRaggedAssay,GRanges}(x, i, j, ..., drop = TRUE)

\S4method{[}{RangedRaggedAssay,ANY}(x, i, j, ..., drop = TRUE)

\S4method{subset}{MultiAssayExperiment}(x, indicator, method = NULL,
  drop = FALSE, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{MultiAssayExperiment}} object}

\item{...}{Additional arguments to pass to SubsetByOverlaps when subsetting by rownames}

\item{drop}{logical (default FALSE) whether to coerce lowest possible dimension after subsetting}

\item{indicator}{A \code{logical} or \code{character} vector or \code{MultiAssayView} class object to use for subsetting}

\item{method}{A \code{character} vector of length one designating to subset either by colnames, rownames, or assays}
}
\value{
A \code{\linkS4class{Elist}} object.

Any type of object describing the metadata

An \code{integer}

A character vector of experiment names
}
\description{
An integrative MultiAssay class for experiment data

Generic Accessor Functions

Generic Acessor Functions

Length of Experiment List

Subset MultiAssayExperiment object
\code{subset} returns a subsetted \code{\linkS4class{MultiAssayExperiment}} object
}
\section{Methods (by generic)}{
\itemize{
\item \code{sampleMap}: Access sampleMap slot from MultiAssayExperiment

\item \code{Elist}: Access Elist class from MultiAssayExperiment

\item \code{masterPheno}: Access masterPheno slot from MultiAssayExperiment

\item \code{metadata}: Access metadata slot from MultiAssayExperiment

\item \code{length}: Get the length of Elist

\item \code{names}: Get the names of the Elist

\item \code{rownames}: Get all the rownames for a MultiAssayExperiment

\item \code{colnames}: Get all the colnames for a MultiAssayExperiment

\item \code{assay}: Get the raw data from a MultiAssayExperiment as a list

\item \code{[}: Subset RangedRaggedAssay using GRanges object

\item \code{[}: Subset RangedRaggedAssay using character vector

\item \code{subset}: export subset
}}
\section{Slots}{

\describe{
\item{\code{Elist}}{A \code{\linkS4class{Elist}} class object for each assay dataset.}

\item{\code{masterPheno}}{A \code{DataFrame} of all clinical data available across experiments.}

\item{\code{sampleMap}}{A \code{DataFrame} of translatable identifiers of samples and participants.}

\item{\code{metadata}}{Additional data describing the \code{\link{MultiAssayExperiment}} object.}

\item{\code{drops}}{A metadata \code{list} of dropped information.}
}}

