% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-methods.R
\docType{methods}
\name{MultiAssayExperiment-helpers}
\alias{MultiAssayExperiment-helpers}
\alias{intersectRows}
\alias{intersectColumns}
\alias{mergeReplicates}
\alias{duplicated}
\alias{intersectColumns}
\alias{duplicated,MultiAssayExperiment-method}
\alias{mergeReplicates}
\alias{mergeReplicates,MultiAssayExperiment-method}
\alias{mergeReplicates,ANY-method}
\title{A group of helper functions for manipulating and cleaning a
MultiAssayExperiment}
\usage{
intersectRows(x)

intersectColumns(x)

\S4method{duplicated}{MultiAssayExperiment}(x, incomparables = FALSE, ...)

mergeReplicates(x, replicates = list(), simplify = BiocGenerics::mean, ...)

\S4method{mergeReplicates}{MultiAssayExperiment}(x, replicates = list(),
  simplify = BiocGenerics::mean, ...)

\S4method{mergeReplicates}{ANY}(x, replicates = list(),
  simplify = BiocGenerics::mean, ...)
}
\arguments{
\item{x}{A MultiAssayExperiment}

\item{incomparables}{unused argument}

\item{...}{mergeReplicates: Additional arguments passed to the
\code{simplify} function}

\item{replicates}{A list of \linkS4class{LogicalList}s
indicating multiple / duplicate entries for each biological unit, see the
\code{duplicated} output}

\item{simplify}{A function for merging repeat measurements in experiments
as indicated by replicates for \code{MultiAssayExperiment}}
}
\description{
A set of helper functions were created to help clean and
manipulate a MultiAssayExperiment object.

\itemize{
    \item complete.cases: Returns a logical vector corresponding to 'colData'
    rows that have data across all experiments
    \item duplicated: Returns a 'list' of 'LogicalList's that indicate
    what measurements originate from the same biological unit
    \item intersectRows: Takes all common rows across experiments,
    excludes experiments with empty rownames
    \item intersectColumns: A wrapper for
    \link[=complete.cases,MultiAssayExperiment-method]{complete.cases} to
    return a MultiAssayExperiment with only those biological units that have
    measurements across all experiments
    \item mergeReplicates: A function that combines duplicated / repeated
    measurements across all experiments and is guided by the duplicated
    return value
}

The \code{mergeReplicates} function is a house-keeping method
for a \code{MultiAssayExperiment} where only \code{complete.cases} are
returned, replicate measurements are averaged (by default), and columns are
aligned by the row order in \code{colData}.

The "ANY" method consolidates duplicate measurements for rectangular data
structures, returns object of the same class (endomorphic)
}
