% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-helpers.R
\docType{methods}
\name{MultiAssayExperiment-helpers}
\alias{MultiAssayExperiment-helpers}
\alias{intersectRows}
\alias{intersectColumns}
\alias{mergeReplicates}
\alias{duplicated}
\alias{intersectColumns}
\alias{duplicated,MultiAssayExperiment-method}
\alias{mergeReplicates}
\alias{mergeReplicates,MultiAssayExperiment-method}
\alias{mergeReplicates,ANY-method}
\alias{longFormat}
\alias{longFormat,ANY-method}
\alias{longFormat,ExperimentList-method}
\alias{longFormat,MultiAssayExperiment-method}
\alias{wideFormat}
\alias{wideFormat,MultiAssayExperiment-method}
\alias{wideFormat,ANY-method}
\title{A group of helper functions for manipulating and cleaning a
MultiAssayExperiment}
\usage{
intersectRows(x)

intersectColumns(x)

\S4method{duplicated}{MultiAssayExperiment}(x, incomparables = FALSE, ...)

mergeReplicates(x, replicates = list(), simplify = BiocGenerics::mean, ...)

\S4method{mergeReplicates}{MultiAssayExperiment}(x, replicates = list(),
  simplify = BiocGenerics::mean, ...)

\S4method{mergeReplicates}{ANY}(x, replicates = list(),
  simplify = BiocGenerics::mean, ...)

longFormat(object, ...)

\S4method{longFormat}{ANY}(object, ...)

\S4method{longFormat}{ExperimentList}(object, ...)

\S4method{longFormat}{MultiAssayExperiment}(object, colDataCols = NULL, ...)

wideFormat(object, ...)

\S4method{wideFormat}{MultiAssayExperiment}(object, colDataCols = NULL,
  key = NULL, ...)

\S4method{wideFormat}{ANY}(object, ...)
}
\arguments{
\item{x}{A MultiAssayExperiment}

\item{incomparables}{unused argument}

\item{...}{Additional arguments. See details.}

\item{replicates}{A list of \linkS4class{LogicalList}s
indicating multiple / duplicate entries for each biological unit, see the
\code{duplicated} output}

\item{simplify}{A function for merging repeat measurements in experiments
as indicated by replicates for \code{MultiAssayExperiment}}

\item{object}{Any supported class object}

\item{colDataCols}{selected colData columns to include in the output}

\item{key}{name of column whose values will used as variables in
the wide dataset from \link[tidyr]{spread}. If none are specified, assay,
rowname, and colname will be combined}
}
\description{
A set of helper functions were created to help clean and
manipulate a MultiAssayExperiment object. \code{intersectRows} also works
for \code{ExperimentList} objects.

\itemize{
    \item complete.cases: Returns a logical vector corresponding to 'colData'
    rows that have data across all experiments
    \item duplicated: Returns a 'list' of 'LogicalList's that indicate
    what measurements originate from the same biological unit
    \item intersectRows: Takes all common rows across experiments,
    excludes experiments with empty rownames
    \item intersectColumns: A wrapper for
    \link[=complete.cases,MultiAssayExperiment-method]{complete.cases} to
    return a MultiAssayExperiment with only those biological units that have
    measurements across all experiments
    \item mergeReplicates: A function that combines duplicated / repeated
    measurements across all experiments and is guided by the duplicated
    return value
    \item longFormat: A \code{MultiAssayExperiment} method that
    returns a small and skinny \link{DataFrame}. The \code{colDataCols}
    arguments allows the user to append \code{colData} columns to the data.
    \item wideFormat: A function to return a wide \link{DataFrame} where
    each row represents an observation. Optional \code{colDataCols} can be
    added when using a \code{MultiAssayExperiment}.
}
}
\details{
For the \code{duplicated} function, the \code{incomparables} and
ellipsis \code{\ldots} arguments are not used.

The \code{mergeReplicates} function is a house-keeping method
for a \code{MultiAssayExperiment} where only \code{complete.cases} are
returned, replicate measurements are averaged (by default), and columns are
aligned by the row order in \code{colData}. Additional arguments can be
passed on to the \code{simplify} function.

The \code{mergeReplicates} "ANY" method consolidates duplicate
measurements for rectangular data structures, returns object of the same
class (endomorphic). The ellipsis or \code{\ldots} argument allows the
user to provide additional arguments to the \code{simplify} functional
argument.

The \code{longFormat} "ANY" class method, works with classes such as
\link{ExpressionSet} and \link{SummarizedExperiment} as well as \code{matrix}
to provide a consistent long and skinny \link{DataFrame}.
}
\section{mergeReplicates}{

The \code{mergeReplicates} function makes use of the output from
\code{duplicated} which will point out the duplicate measurements by
biological unit in the \code{MultiAssayExperiment}. This function will return
a \code{MultiAssayExperiment} with merged replicates.
}

\section{longFormat}{

The longFormat method takes data from the \code{\link{ExperimentList}}
in a \code{\link{MultiAssayExperiment}} and returns a uniform
\code{\link{DataFrame}}. The resulting DataFrame has columns indicating
primary, rowname, colname and value. This method can optionally include
colData columns with the \code{colDataCols} argument
(\code{MultiAssayExperiment} method only).
}

\section{wideFormat}{

The \code{wideFormat} \code{MultiAssayExperiment} method returns standardized
wide \link{DataFrame} where each row represents an observation or biological
unit as represented in \code{colData}. Optionally, \code{colData} columns
can be added to the data output. The \code{wideFormat} method for an
\code{ExperimentList} returns a list of wideFormat \code{DataFrames}. The
"ANY" method returns a wide format \code{DataFrame}.
}

