% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment.R
\docType{package}
\name{MultiAssayExperiment-package}
\alias{MultiAssayExperiment}
\alias{MultiAssayExperiment-package}
\title{Create an integrative multiassay container}
\description{
MultiAssayExperiment allows the manipulation of related multiassay datasets
}
\details{
An integrative package for coordinated representation of multiple experiments
on partially overlapping samples, with associated metadata at the level of
an entire study and at the level of the "biological unit". The biological
unit may be a patient, plant, yeast strain, etc. This package is designed
around the following hierarchy of information: study, experiments, samples
}

