% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-pkg.R
\docType{package}
\name{MultiAssayExperiment-package}
\title{MultiAssayExperiment: Build an integrative multi-assay container}
\description{
MultiAssayExperiment allows the manipulation of related multiassay
datasets with partially overlapping samples, associated metadata at
the level of an entire study, and at the level of the "biological unit".
The biological unit may be a patient, plant, yeast strain, etc.
}
\details{
The package hierarchy of information:
\itemize{
\item study
\item experiments
\item samples
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vjcitn/MultiAssayExperiment/wiki/MultiAssayExperiment-
API}
  \item Report bugs at \url{https://github.com/vjcitn/MultiAssayExperiment/issues}
}

}

