% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-pkg.R, R/MultiAssayExperiment.R
\docType{package}
\name{MultiAssayExperiment}
\alias{MultiAssayExperiment}
\alias{MultiAssayExperiment-package}
\title{MultiAssayExperiment: Build an integrative multi-assay container}
\usage{
MultiAssayExperiment(ExperimentList = list(),
  pData = S4Vectors::DataFrame(), sampleMap = S4Vectors::DataFrame(),
  drops = list())
}
\arguments{
\item{ExperimentList}{A \code{list} of all combined experiments}

\item{pData}{A \code{\link[S4Vectors]{DataFrame}} of the phenotype
data for all participants}

\item{sampleMap}{A \code{DataFrame} of sample identifiers, assay samples,
and assay names}

\item{drops}{A \code{list} of unmatched information
(included after subsetting)}
}
\value{
A \code{MultiAssayExperiment} data object that stores experiment
and phenotype data
}
\description{
MultiAssayExperiment allows the manipulation of related multiassay 
datasets with partially overlapping samples, associated metadata at 
the level of an entire study, and at the level of the "biological unit".
The biological unit may be a patient, plant, yeast strain, etc.

This function combines multiple data elements from the different hierarchies
of data (study, experiments, and samples)
}
\details{
The package hierarchy of information: 
\itemize{
\item study
\item experiments
\item samples
}
}
\examples{
## Run the example ExperimentList
example("ExperimentList")

## Load example GRangesList object
example("RangedRaggedAssay")

## Add the RangedRaggedAssay to the list
ExpList <- c(ExpList, myRRA)
names(ExpList)[3] <- "CNVgistic"

## Run the sample map example
example("sampleMap")

## Create an example phenotype data
primary <- data.frame(sex = c("M", "F", "M", "F"),
                       age = 38:41,
                       row.names = c("Jack", "Jill", "Bob", "Barbara"))

## Create a MultiAssayExperiment instance
myMultiAssayExperiment <- MultiAssayExperiment(ExperimentList = ExpList,
                                               pData = primary,
                                               sampleMap = mySampleMap)
}
\seealso{
MultiAssayExperiment-class
}

