% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiAssayExperiment.R
\name{MultiAssayExperiment}
\alias{MultiAssayExperiment}
\title{Create a MultiAssayExperiment object
\code{MultiAssayExperiment} returns a \code{\linkS4class{MultiAssayExperiment}} object}
\usage{
MultiAssayExperiment(explist = list(), masterPheno = data.frame(),
  sampleMap = list(), drops = list())
}
\arguments{
\item{explist}{A \code{list} of all combined experiments}

\item{masterPheno}{A \code{data.frame} of the phenotype data for all participants.}

\item{sampleMap}{A \code{data.frame} of sample identifiers, assay samples, and assay names.}

\item{drop}{Logical (default FALSE) parameter for dropping samples with unmatched phenotype data.}
}
\value{
A \code{MultiAssayExperiment} data object that stores experiment and phenotype data.
}
\description{
This function combines multiple data sources specific to one disease by matching samples.
}

