% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment.R
\name{MultiAssayExperiment}
\alias{MultiAssayExperiment}
\title{Construct a \code{MultiAssayExperiment} object}
\usage{
MultiAssayExperiment(experiments = ExperimentList(),
  pData = S4Vectors::DataFrame(), sampleMap = S4Vectors::DataFrame(),
  metadata = NULL, drops = list())
}
\arguments{
\item{experiments}{A \code{list} or \link{ExperimentList} of all
combined experiments}

\item{pData}{A \code{\link[S4Vectors]{DataFrame}} or \code{data.frame} of
the phenotype data for all participants}

\item{sampleMap}{A \code{DataFrame} or \code{data.frame} of assay names,
sample identifiers, and colname samples}

\item{metadata}{An optional argument of "ANY" class (usually list) for
content describing the experiments}

\item{drops}{A \code{list} of unmatched information
(included after subsetting)}
}
\value{
A \code{MultiAssayExperiment} object that can store
experiment and phenotype data
}
\description{
The constructor function for the \link{MultiAssayExperiment-class} combines
multiple data elements from the different hierarchies of data
(study, experiments, and samples). It can create instances where neither
a \code{sampleMap} or a \code{pData} set is provided. Please see the
MultiAssayExperiment API documentation for more information by running the
\code{API} function.
}
\examples{
## Run the example ExperimentList
example("ExperimentList")

## Load example GRangesList object
example("RangedRaggedAssay")

## Add the RangedRaggedAssay to the list
ExpList <- c(ExpList, CNVgistic = myRRA)

## Create sample maps for each experiment
exprmap <- data.frame(
    primary = c("Jack", "Jill", "Barbara", "Bob"),
    colname = c("array1", "array2", "array3", "array4"),
    stringsAsFactors = FALSE)

methylmap <- data.frame(
    primary = c("Jack", "Jack", "Jill", "Barbara", "Bob"),
    colname = c("methyl1", "methyl2", "methyl3", "methyl4", "methyl5"),
    stringsAsFactors = FALSE)

rangemap <- data.frame(
    primary = c("Jack", "Jill", "Jill"),
    colname = c("snparray1", "snparray2", "snparray3"),
    stringsAsFactors = FALSE)

rnamap <- data.frame(
    primary = c("Jack", "Jill", "Bob", "Barbara"),
    colname = c("samparray1", "samparray2", "samparray3",
                               "samparray4"),
    stringsAsFactors = FALSE)

## Combine as a named list and convert to a DataFrame
mylist <- list(Affy = exprmap, Methyl450k = methylmap,
               CNVgistic = rangemap, RNASeqGene = rnamap)

## Create a sampleMap
mySampleMap <- listToMap(mylist)
## Create an example phenotype data
pDat <- data.frame(sex = c("M", "F", "M", "F"),
                       age = 38:41,
                       row.names = c("Jack", "Jill", "Bob", "Barbara"))

## Create a MultiAssayExperiment instance
myMultiAssayExperiment <- MultiAssayExperiment(experiments = ExpList,
                                               pData = pDat,
                                               sampleMap = mySampleMap)
}
\seealso{
\link{MultiAssayExperiment-class}
}
