% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment.R
\name{MultiAssayExperiment}
\alias{MultiAssayExperiment}
\title{Create a MultiAssayExperiment object}
\usage{
MultiAssayExperiment(Elist = list(), pData = S4Vectors::DataFrame(),
  sampleMap = S4Vectors::DataFrame(), drops = list())
}
\arguments{
\item{Elist}{A \code{list} of all combined experiments}

\item{pData}{A \code{\link[S4Vectors]{DataFrame-class}} of the phenotype
data for all participants}

\item{sampleMap}{A \code{DataFrame} of sample identifiers, assay samples,
and assay names}

\item{drops}{A \code{list} of unmatched information
(included after subsetting)}
}
\value{
A \code{MultiAssayExperiment} data object that stores experiment
and phenotype data
}
\description{
This function combines multiple data elements from the different hierarchies 
of data (study, experiments, and samples)
}

