% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayView-class.R, R/biocMultiAssay-methods.R
\docType{class}
\name{MultiAssayView-class}
\alias{.MultiAssayView}
\alias{MultiAssayView}
\alias{MultiAssayView-class}
\alias{length,MultiAssayView-method}
\alias{names,MultiAssayView-method}
\title{An class used for staging a subset operation}
\usage{
\S4method{names}{MultiAssayView}(x)

\S4method{length}{MultiAssayView}(x)
}
\arguments{
\item{x}{A \code{\linkS4class{MultiAssayView}} class object}
}
\value{
A character vector of experiment names
}
\description{
An class used for staging a subset operation
}
\section{Methods (by generic)}{
\itemize{
\item \code{names}: Get the names from the kept elements in the Elist

\item \code{length}: Get the number of assays from kept slot
}}
\section{Slots}{

\describe{
\item{\code{subject}}{A MultiAssayExperiment instance}

\item{\code{rowindex}}{An \code{IntegerList} indexing rows of subject
retained for subsequent operations}

\item{\code{colindex}}{An \code{IntegerList} indexing columns of sujbect
retained for subsequent operations}

\item{\code{assayindex}}{An \code{IntegerList} indexing assays of subject
retained for subsequent operations}
}}

