% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-methods.R, R/MultiAssayView-class.R
\docType{methods}
\name{names,MultiAssayView-method}
\alias{.MultiAssayView}
\alias{MultiAssayView}
\alias{MultiAssayView-class}
\alias{[,MultiAssayView,ANY,ANY,ANY-method}
\alias{colnames,MultiAssayView-method}
\alias{length,MultiAssayView-method}
\alias{names,MultiAssayView-method}
\alias{rownames,MultiAssayView-method}
\alias{show,MultiAssayView-method}
\title{A class used for staging a subset operation}
\usage{
\S4method{names}{MultiAssayView}(x)

\S4method{length}{MultiAssayView}(x)

\S4method{rownames}{MultiAssayView}(x)

\S4method{colnames}{MultiAssayView}(x)

\S4method{[}{MultiAssayView,ANY,ANY,ANY}(x, i, j, k, ..., drop = TRUE)

\S4method{show}{MultiAssayView}(object)
}
\arguments{
\item{x}{A \code{MultiAssayView} class object}

\item{i}{A \code{character} vector for subsetting rownames}

\item{j}{A \code{character} vector for subsetting colnames}

\item{k}{A \code{character} vector for subsetting assays}

\item{...}{Additional parameters}

\item{drop}{logical (default TRUE) whether to drop empty assay elements}

\item{object}{A \code{MultiAssayView} class object}
}
\value{
A \code{MultiAssayView} object
}
\description{
Use a \code{MultiAssayView} class to initialize a subsetting operation
of a \code{\linkS4class{MultiAssayExperiment}} class object
}
\section{Methods (by generic)}{
\itemize{
\item \code{names}: Get a \code{character} vector of experiment names

\item \code{length}: Get the number of assays in the
\code{MultiAssayExperiment} instance

\item \code{rownames}: Get a CharacterList of rownames

\item \code{colnames}: Get a CharacterList of colnames

\item \code{[}: Subset MultiAssayView dimensions with a 
\code{character} vector

\item \code{show}: Show method for \code{MultiAssayView}
}}
\section{Slots}{

\describe{
\item{\code{subject}}{A \code{\linkS4class{MultiAssayExperiment}} instance}

\item{\code{rowindex}}{An \code{IntegerList} indexing rows of subject
retained for subsequent operations}

\item{\code{colindex}}{An \code{IntegerList} indexing columns of sujbect
retained for subsequent operations}

\item{\code{assayindex}}{An \code{integer} indexing assays of subject
retained for subsequent operations}
}}

