% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayView-class.R
\name{MultiAssayView}
\alias{MultiAssayView}
\title{A parallel representation for manipulating a MultiAssayExperiment}
\usage{
MultiAssayView(subject)
}
\arguments{
\item{subject}{A \code{MultiAssayExperiment} class object}
}
\value{
A \code{MultiAssayView} representation of a loaded
\code{MultiAssayExperiment} object.
}
\description{
The purpose of the \code{MultiAssayView} class is to provide a
lightweight and memory efficient representation of a 
\code{MultiAssayExperiment} class object.
}
\examples{
example("MultiAssayExperiment")
MultiAssayView(myMultiAssayExperiment)

}

