% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayView.R
\name{MultiAssayView}
\alias{MultiAssayView}
\title{MultiAssayView operation for colnames, rownames, or assay}
\usage{
MultiAssayView(MultiAssay, identifier, method = character(), ...)
}
\arguments{
\item{MultiAssay}{A \code{\linkS4class{MultiAssayExperiment}}}

\item{identifier}{Either a \code{character}, \code{numeric} or \code{logical} vector identifying targets}

\item{method}{Prepare/View for subsetting using colnames, rownames or assays.}

\item{...}{Additional arguments passed to the findOverlaps function}
}
\value{
A \code{\linkS4class{MultiAssayView}} class object for subsequent subsetting
}
\description{
MultiAssayView operation for colnames, rownames, or assay
}

