% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\name{pData}
\alias{pData}
\title{Generic Accessor Functions}
\usage{
pData(x)
}
\arguments{
\item{x}{A \code{\link{MultiAssayExperiment}} object}
}
\value{
A \code{DataFrame} object
}
\description{
Generic Accessor Functions
}

