% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepMultiAssay.R
\name{PrepMultiAssay}
\alias{PrepMultiAssay}
\title{Prepare a MultiAssayExperiment instance}
\usage{
PrepMultiAssay(Elist, pData, sampleMap)
}
\arguments{
\item{Elist}{A \code{list} of all combined experiments}

\item{pData}{A \code{\link[S4Vectors]{DataFrame-class}} of the phenotype
data for all participants}

\item{sampleMap}{A \code{DataFrame} of sample identifiers, assay samples,
and assay names}
}
\value{
Messages and suggestions for creating a MultiAssayExperiment
}
\description{
The purpose of this helper function is to faciltate the creation of a
MultiAssayExperiment object by detecting any inconsistencies with all types
of names in either the Elist or the pData.
}

