% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaggedRangedAssay-class.R
\docType{class}
\name{RaggedRangedAssay-class}
\alias{RaggedRangedAssay,GRangesList-method}
\alias{RaggedRangedAssay-class}
\title{An extension of the GRangesList class}
\usage{
\S4method{RaggedRangedAssay}{GRangesList}(...)
}
\arguments{
\item{...}{A \code{GRangesList} object}
}
\value{
A \code{\linkS4class{RaggedRangedAssay}} class object
}
\description{
An extension of the GRangesList class

Convert GRangesList to RaggedRangedAssay
}
\section{Methods (by generic)}{
\itemize{
\item \code{RaggedRangedAssay}: Convert a GRangesList to a RaggedRangedAssay
}}

