% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\docType{class}
\name{RangedRaggedAssay-class}
\alias{RangedRaggedAssay-class}
\alias{.RangedRaggedAssay}
\title{RangedRaggedAssay}
\arguments{
\item{x}{A \code{list}, \code{GRanges} or \code{GRangesList} object}
}
\description{
This class has been deprecated in place for the RaggedExperiment class.
Please see that package for representing mutation and copy number data.
}
