% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangedRaggedAssay-class.R, R/biocMultiAssay-methods.R
\docType{class}
\name{RangedRaggedAssay-class}
\alias{.RangedRaggedAssay}
\alias{RangedRaggedAssay-class}
\alias{[,RangedRaggedAssay,ANY-method}
\alias{[,RangedRaggedAssay,GRanges-method}
\alias{assay,RangedRaggedAssay,ANY-method}
\alias{colnames,RangedRaggedAssay-method}
\alias{getHits,RangedRaggedAssay,character-method}
\alias{ncol,RangedRaggedAssay-method}
\alias{nrow,RangedRaggedAssay-method}
\alias{rownames,RangedRaggedAssay-method}
\title{An extension of the GRangesList class}
\usage{
\S4method{ncol}{RangedRaggedAssay}(x)

\S4method{nrow}{RangedRaggedAssay}(x)

\S4method{[}{RangedRaggedAssay,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{RangedRaggedAssay,GRanges}(x, i, j, ..., drop = TRUE)

\S4method{rownames}{RangedRaggedAssay}(x)

\S4method{colnames}{RangedRaggedAssay}(x)

\S4method{assay}{RangedRaggedAssay,ANY}(x)

\S4method{getHits}{RangedRaggedAssay,character}(subject, query, ...)
}
\arguments{
\item{x}{A \code{\link{RangedRaggedAssay}} class}

\item{i}{Either a \code{character} or \code{GRanges} class object 
to subset by rows}

\item{j}{Either a \code{character}, \code{numeric}, or \code{logical} 
type for selecting columns (\code{\link[GenomicRanges]{GRangesList}} method)}

\item{...}{Any additional arguments passed on to subsetByOverlaps}

\item{drop}{logical (defalut TRUE) whether to drop empty columns}

\item{subject}{A \code{RangedRaggedAssay} class object}

\item{query}{A \code{character} class for searching hits}
}
\value{
A \code{\link{RangedRaggedAssay}} class object
}
\description{
An extension of the GRangesList class

Subsetting a RangedRaggedAssay can be done using either rownames and column
names
}
\section{Methods (by generic)}{
\itemize{
\item \code{ncol}: Get the column length of a RangedRaggedAssay

\item \code{nrow}: Get the row length of a RangedRaggedAssay

\item \code{[}: Subset a \code{RangedRaggedAssay} with either 
\code{chracter}, \code{numeric}, or \code{logical}

\item \code{[}: Subset a \code{RangedRaggedAssay} using a
\code{GRanges} class object

\item \code{rownames}: Get feature names from a RangedRaggedAssay

\item \code{colnames}: Get sample names from a RangedRaggedAssay

\item \code{assay}: Get experiment metadata from a 
RangedRaggedAssay

\item \code{getHits}: Find matching features by character in a 
RangedRaggedAssay
}}
\seealso{
\code{\link[IRanges]{findOverlaps-methods}}
}

