% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangedRaggedAssay-class.R, R/MultiAssayExperiment-methods.R
\docType{class}
\name{RangedRaggedAssay-class}
\alias{.RangedRaggedAssay}
\alias{RangedRaggedAssay-class}
\alias{[,RangedRaggedAssay,ANY,ANY,ANY-method}
\alias{[,RangedRaggedAssay,GRanges,ANY,ANY-method}
\alias{colnames,RangedRaggedAssay-method}
\alias{colnames<-,RangedRaggedAssay,character-method}
\alias{dim,RangedRaggedAssay-method}
\alias{getHits,RangedRaggedAssay,character-method}
\alias{ncol,RangedRaggedAssay-method}
\alias{nrow,RangedRaggedAssay-method}
\alias{rownames,RangedRaggedAssay-method}
\title{An extension of the GRangesList class}
\usage{
\S4method{[}{RangedRaggedAssay,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{RangedRaggedAssay,GRanges,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{dim}{RangedRaggedAssay}(x)

\S4method{ncol}{RangedRaggedAssay}(x)

\S4method{nrow}{RangedRaggedAssay}(x)

\S4method{rownames}{RangedRaggedAssay}(x)

\S4method{colnames}{RangedRaggedAssay}(x)

\S4method{colnames}{RangedRaggedAssay,character}(x) <- value

\S4method{getHits}{RangedRaggedAssay,character}(subject, query, ...)
}
\arguments{
\item{x}{A \code{\link{RangedRaggedAssay}} class}

\item{i}{Either a \code{character} or \code{GRanges} class object
to subset by rows}

\item{j}{Either a \code{character}, \code{numeric}, or \code{logical} 
type for selecting columns (\code{\link[GenomicRanges]{GRangesList}} method)}

\item{...}{Any additional arguments passed on to subsetByOverlaps}

\item{drop}{logical (default TRUE) whether to drop empty columns}

\item{value}{A \code{character} vector representing column or sample names}

\item{subject}{A \code{RangedRaggedAssay} class object}

\item{query}{A \code{character} class for searching hits}
}
\value{
A \code{\link{RangedRaggedAssay}} class object
}
\description{
An extension of the GRangesList class

Subsetting a RangedRaggedAssay can be done using either rownames and column
names
}
\section{Methods (by generic)}{
\itemize{
\item \code{[}: Subset a \code{RangedRaggedAssay} with either 
\code{chracter}, \code{numeric}, or \code{logical}

\item \code{[}: Subset a \code{RangedRaggedAssay} using a
\code{GRanges} class object

\item \code{dim}: Obtain dimension lengths of a
\code{RangedRaggedAssay} class object

\item \code{ncol}: Get the column length of a
\code{RangedRaggedAssay} class object

\item \code{nrow}: Get the row length of a
\code{RangedRaggedAssay} class object

\item \code{rownames}: Get feature names from a
\code{RangedRaggedAssay}

\item \code{colnames}: Get sample names from a
\code{RangedRaggedAssay}

\item \code{colnames<-}: value: A modified \code{RangedRaggedAssay}
object

\item \code{getHits}: Find matching features by \code{character}
in a \code{RangedRaggedAssay}
}}
\seealso{
\code{\link[IRanges]{findOverlaps-methods}}
}

