% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangedRaggedAssay-class.R,
%   R/MultiAssayExperiment-methods.R
\docType{class}
\name{RangedRaggedAssay-class}
\alias{RangedRaggedAssay-class}
\alias{.RangedRaggedAssay}
\alias{[,RangedRaggedAssay,ANY,ANY,ANY-method}
\alias{[,RangedRaggedAssay,ANY-method}
\alias{[,RangedRaggedAssay,GRanges,ANY,ANY-method}
\alias{[,RangedRaggedAssay,GRanges-method}
\alias{dim,RangedRaggedAssay-method}
\alias{dimnames,RangedRaggedAssay-method}
\alias{dimnames<-,RangedRaggedAssay,list-method}
\alias{disjoin,RangedRaggedAssay-method}
\alias{show,RangedRaggedAssay-method}
\alias{reduce,RangedRaggedAssay-method}
\title{An extension of the GRangesList class}
\usage{
\S4method{[}{RangedRaggedAssay,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{RangedRaggedAssay,GRanges,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{dim}{RangedRaggedAssay}(x)

\S4method{dimnames}{RangedRaggedAssay}(x)

\S4method{dimnames}{RangedRaggedAssay,list}(x) <- value

\S4method{disjoin}{RangedRaggedAssay}(x, mcolname = NULL, FUN = mean, ...)

\S4method{show}{RangedRaggedAssay}(object)

\S4method{reduce}{RangedRaggedAssay}(x, drop.empty.ranges = FALSE,
  replicates = NULL, combine = rowMeans, vectorized = TRUE,
  mcolname = NULL, ...)
}
\arguments{
\item{x}{A \code{\link{RangedRaggedAssay}} class}

\item{i}{Either a \code{character} or \code{GRanges} class object
to subset by rows}

\item{j}{Either a \code{character}, \code{numeric}, or \code{logical}
type for selecting columns (\code{\link[GenomicRanges]{GRangesList}} method)}

\item{...}{Additional arguments. See details for more information.}

\item{drop}{logical (default TRUE) whether to drop empty columns}

\item{value}{A \code{list} object of row and column names}

\item{mcolname}{A single character string indicating metadata column to use
for summaries}

\item{FUN}{A function for summarizing non-disjoint ranges (default mean)}

\item{object}{A \code{RangedRaggedAssay} class object}

\item{drop.empty.ranges}{unused argument}

\item{replicates}{reduce: A logical list where each element represents a
sample and a vector of repeated experiments for the sample (default NULL)}

\item{combine}{A function for consolidating columns in the matrix
representation of the data (default rowMeans)}

\item{vectorized}{logical (default TRUE) whether the \code{combine} function
is vectorized, optimized for working down the vector pairs}
}
\value{
A \code{\link{RangedRaggedAssay}} class object
}
\description{
An extension of the GRangesList class

Subsetting a RangedRaggedAssay can be done using either rownames and column
names
}
\details{
The \code{\ldots} argument allows the user to specify arguments in the
\code{\link{subsetByOverlaps}} function. When calling the \code{reduce}
method, the additional arguments correspond to those in either the
\code{assay} method or the \code{reduce} method. The \code{reduce} arguments
include a function for applying over the rows (\code{combine}) and a
\code{vectorized} argument which indicates whether the given function is
vectorized or not.
}
\section{Methods (by generic)}{
\itemize{
\item \code{[}: Subset a \code{RangedRaggedAssay} with either
\code{chracter}, \code{numeric}, or \code{logical}

\item \code{[}: Subset a \code{RangedRaggedAssay} using a
\code{GRanges} class object

\item \code{dim}: Obtain dimension lengths of a
\code{RangedRaggedAssay} class object

\item \code{dimnames}: Get dimension names
for a \code{RangedRaggedAssay}

\item \code{dimnames<-}: value: A modified \code{RangedRaggedAssay}
object

\item \code{disjoin}: Separate non-disjoint ranges and merge
with function

\item \code{show}: show method for
the \code{RangedRaggedAssay} class

\item \code{reduce}: Use metadata column to produce a matrix which
can then be merged across replicates.
}}

\examples{
## Load sample MultiAssayExperiment Copy Number data
dataFile <- system.file("extdata", "hnscSample.rds",
    package = "MultiAssayExperiment")

hnscSample <- readRDS(dataFile)

## assay method for a RangedRaggedAssay
assay(hnscSample[[1L]], mcolname = "Segment_Mean")[1:5, 1:3]

hnscSample[[2]] <- disjoin(hnscSample[[2L]])

matrices <- assay(hnscSample, mcolname = "Segment_Mean")

length(matrices)
class(matrices)
}
\seealso{
\code{\link[IRanges]{findOverlaps-methods}}

\link{assay,RangedRaggedAssay,missing-method}
}
