% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangedRaggedAssay-class.R,
%   R/MultiAssayExperiment-methods.R
\docType{class}
\name{RangedRaggedAssay-class}
\alias{RangedRaggedAssay-class}
\alias{[,RangedRaggedAssay,ANY,ANY,ANY-method}
\alias{[,RangedRaggedAssay,ANY-method}
\alias{[,RangedRaggedAssay,GRanges,ANY,ANY-method}
\alias{[,RangedRaggedAssay,GRanges-method}
\alias{dim,RangedRaggedAssay-method}
\alias{dimnames,RangedRaggedAssay-method}
\alias{dimnames<-,RangedRaggedAssay,list-method}
\alias{disjoin,RangedRaggedAssay-method}
\alias{show,RangedRaggedAssay-method}
\alias{mergeReplicates,RangedRaggedAssay-method}
\title{An extension of the GRangesList class}
\usage{
\S4method{[}{RangedRaggedAssay,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{RangedRaggedAssay,GRanges,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{dim}{RangedRaggedAssay}(x)

\S4method{dimnames}{RangedRaggedAssay}(x)

\S4method{dimnames}{RangedRaggedAssay,list}(x) <- value

\S4method{disjoin}{RangedRaggedAssay}(x, mcolname = NULL,
  simplify = BiocGenerics::mean, ...)

\S4method{show}{RangedRaggedAssay}(object)

\S4method{mergeReplicates}{RangedRaggedAssay}(object, replicates = list(),
  simplify = BiocGenerics::mean, mcolname = NULL, ...)
}
\arguments{
\item{x}{A \code{\link{RangedRaggedAssay}} class}

\item{i}{Either a \code{character} or \code{GRanges} class object
to subset by rows}

\item{j}{Either a \code{character}, \code{numeric}, or \code{logical}
type for selecting columns (\code{\link[GenomicRanges]{GRangesList}} method)}

\item{...}{Additional arguments. See details for more information.}

\item{drop}{logical (default TRUE) whether to drop empty columns}

\item{value}{A \code{list} object of row and column names}

\item{mcolname}{A single character string indicating metadata column to use
for summaries}

\item{simplify}{A function for combining duplicate measurements (e.g., mean)}

\item{object}{A \code{RangedRaggedAssay} class object}

\item{replicates}{mergeReplicates: A \code{list} or \code{LogicalList} where
each element represents a sample and a vector of repeated measurements for
that biological unit}
}
\value{
A \code{\link{RangedRaggedAssay}} class object
}
\description{
An extension of the GRangesList class

Subsetting a RangedRaggedAssay can be done using either rownames and column
names
}
\details{
The \code{\ldots} argument allows the user to specify arguments in the
\code{\link{subsetByOverlaps}} function. When calling the
\code{mergeReplicates} method, the additional arguments correspond to those
in either the \code{assay} method or the \code{mergeReplicates} method.
The \code{mergeReplicates} arguments include a function for applying over
the rows (\code{combine}) and a \code{vectorized} argument which indicates
whether the given function is vectorized or not.
}
\section{Methods (by generic)}{
\itemize{
\item \code{[}: Subset a \code{RangedRaggedAssay} with either
\code{chracter}, \code{numeric}, or \code{logical}

\item \code{[}: Subset a \code{RangedRaggedAssay} using a
\code{GRanges} class object

\item \code{dim}: Obtain dimension lengths of a
\code{RangedRaggedAssay} class object

\item \code{dimnames}: Get dimension names
for a \code{RangedRaggedAssay}

\item \code{dimnames<-}: value: A modified \code{RangedRaggedAssay}
object

\item \code{disjoin}: Separate non-disjoint ranges and merge
with function

\item \code{show}: show method for
the \code{RangedRaggedAssay} class

\item \code{mergeReplicates}: (deprecated) Use metadata column to produce
a matrix which can then be merged across replicates.
}}

\seealso{
\code{\link[IRanges]{findOverlaps-methods}}

\link{assay,RangedRaggedAssay,missing-method}
}
