% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangedRaggedAssay-class.R
\name{RangedRaggedAssay}
\alias{RangedRaggedAssay}
\title{Create a RangedRaggedAssay}
\usage{
RangedRaggedAssay(x = list())
}
\arguments{
\item{x}{A \code{list}, \code{GRanges} or \code{GRangesList} object}
}
\value{
A \code{\linkS4class{RangedRaggedAssay}} class object
}
\description{
Create a RangedRaggedAssay
}

