% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biocMultiAssay-classes.R
\name{sampleExtractor}
\alias{sampleExtractor}
\title{Sample extractor for eSet, SummarizedExperiment, matrix, and GRangesList}
\usage{
sampleExtractor(x)
}
\arguments{
\item{x}{Either an \code{\linkS4class{ExpressionSet}}, \code{\linkS4class{GRangesList}}, \code{\linkS4class{SummarizedExperiment}} or \code{matrix} class object}
}
\value{
Returns an object of the same class
}
\description{
Sample extractor for eSet, SummarizedExperiment, matrix, and GRangesList
}

