% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\name{sampleMap<-}
\alias{sampleMap<-}
\title{Replace a slot value with a given \code{DataFrame}}
\usage{
sampleMap(x) <- value
}
\arguments{
\item{x}{A \code{MultiAssayExperiment} object}

\item{value}{A \code{DataFrame} object to replace the existing
\code{sampleMap}}
}
\description{
Replace a slot value with a given \code{DataFrame}
}

