% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\name{sampleMap}
\alias{sampleMap}
\title{Show method for MultiAssayExperiment class}
\usage{
sampleMap(x)
}
\arguments{
\item{x}{A \code{\linkS4class{MultiAssayExperiment}} object.}
}
\value{
A \code{"list"} object.
}
\description{
param object A \code{\linkS4class{MultiAssayExperiment}} object.
return Returns a summary of contents for the \code{\linkS4class{MultiAssayExperiment}} class.
exportMethod "show"
Generic Accessor Functions
}

