% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\name{sampleMap}
\alias{sampleMap}
\title{Accessor function for the \code{sampleMap} slot of a
\code{MultiAssayExperiment} object}
\usage{
sampleMap(x)
}
\arguments{
\item{x}{A \code{MultiAssayExperiment} object}
}
\value{
A \code{DataFrame} object of sample relationships across experiments
}
\description{
Accessor function for the \code{sampleMap} slot of a
\code{MultiAssayExperiment} object
}

