% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\name{sampleMap}
\alias{sampleMap}
\title{Show method for MultiAssayExperiment class}
\usage{
sampleMap(x)
}
\arguments{
\item{object}{A \code{\linkS4class{MultiAssayExperiment}}}
}
\value{
Returns a list of contents for the MultiAssayExperiment
}
\description{
Show method for MultiAssayExperiment class
}

