% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\docType{methods}
\name{sampleMap}
\alias{sampleMap}
\alias{sampleMap,MultiAssayExperiment-method}
\title{Show method for MultiAssayExperiment class}
\usage{
sampleMap(x)

\S4method{sampleMap}{MultiAssayExperiment}(x)
}
\arguments{
\item{x}{A \code{\linkS4class{MultiAssayExperiment}} object.}
}
\value{
A \code{"list"} object.
}
\description{
param object A \code{\linkS4class{MultiAssayExperiment}} object.
return Returns a summary of contents for the \code{\linkS4class{MultiAssayExperiment}} class.
exportMethod "show"
Generic Accessor Functions
}
\section{Methods (by class)}{
\itemize{
\item \code{MultiAssayExperiment}: 
}}

