% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\name{sampleMap}
\alias{sampleMap}
\title{Accessor function for the \code{sampleMap} slot of a
\code{MultiAssayExperiment} object}
\usage{
sampleMap(x)
}
\arguments{
\item{x}{A \code{MultiAssayExperiment} object}
}
\value{
A \code{DataFrame} object of sample relationships across experiments
}
\description{
Accessor function for the \code{sampleMap} slot of a
\code{MultiAssayExperiment} object
}
\examples{
## Create sample maps for each experiment
exprmap <- data.frame(
    primary = c("Jack", "Jill", "Barbara", "Bob"),
    colname = c("array1", "array2", "array3", "array4"),
    stringsAsFactors = FALSE)

methylmap <- data.frame(
    primary = c("Jack", "Jack", "Jill", "Barbara", "Bob"),
    colname = c("methyl1", "methyl2", "methyl3", "methyl4", "methyl5"),
    stringsAsFactors = FALSE)

rangemap <- data.frame(primary = c("Jack", "Jill", "Jill"),
    colname = c("snparray1", "snparray2", "snparray3"),
    stringsAsFactors = FALSE)

## Combine as a named list and convert to a DataFrame
mylist <- list(exprmap, methylmap, rangemap)
names(mylist) <- c("Affy", "Methyl450k", "CNVgistic")

## Create a sampleMap
mySampleMap <- listToMap(mylist)
}

