% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-class.R
\name{sampleMap}
\alias{sampleMap}
\title{Generic Accessor Functions}
\usage{
sampleMap(x)
}
\arguments{
\item{x}{A \code{\link{MultiAssayExperiment}} object.}
}
\value{
A \code{list} object.
}
\description{
Generic Accessor Functions
}

