% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elist-class.R
\docType{methods}
\name{show,Elist-method}
\alias{show,Elist-method}
\title{Show method for \code{\linkS4class{Elist}} class}
\usage{
\S4method{show}{Elist}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{Elist}} object.}
}
\value{
Returns a summary of contents for the \code{\linkS4class{Elist}}
class
}
\description{
Show method for \code{\linkS4class{Elist}} class
}

