% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-class.R
\docType{methods}
\name{show,stage-method}
\alias{show,stage-method}
\title{Show method for \code{\linkS4class{stage}} class}
\usage{
\S4method{show}{stage}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{stage}} class object}
}
\value{
A summary of \code{\linkS4class{stage}} class contents
}
\description{
Show method for \code{\linkS4class{stage}} class
}

