% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage-class.R, R/biocMultiAssay-methods.R
\docType{class}
\name{Stage-class}
\alias{Stage-class}
\alias{length,Stage-method}
\alias{names,Stage-method}
\alias{query,Stage-method}
\alias{type,Stage-method}
\title{An identifier class used for staging a subset operation}
\usage{
\S4method{names}{Stage}(x)

\S4method{length}{Stage}(x)

\S4method{type}{Stage}(object)

\S4method{query}{Stage}(object)
}
\arguments{
\item{x}{A \code{\link{Stage}} class object}
}
\value{
A character vector of experiment names
}
\description{
An identifier class used for staging a subset operation

Names of Experiments
}
\section{Methods (by generic)}{
\itemize{
\item \code{names}: Get the names from the kept elements in the Elist

\item \code{length}: Get the number of assays from kept slot

\item \code{type}: Get the staging type (either by samples, features, assays)

\item \code{query}: Get the identifiers used for staging
}}
\section{Slots}{

\describe{
\item{\code{query}}{Any class indicator needed to subset}

\item{\code{keeps}}{A \code{list} indicating valid matches in each assay}

\item{\code{drops}}{A \code{list} of excluded information due to subsetting}

\item{\code{type}}{A \code{character} vector indicating method used to search}
}}

