% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage.R
\name{stage}
\alias{stage}
\title{Stage by Samples, Features, or Assays}
\usage{
stage(MultiAssay, identifier, by = NULL)
}
\arguments{
\item{MultiAssay}{A \code{\linkS4class{MultiAssayExperiment}}}

\item{identifer}{Either a \code{character}, \code{numeric} or \code{logical} vector identifying targets}
}
\value{
A \code{\linkS4class{stage}} class object for subsequent subsetting
}
\description{
Stage by Samples, Features, or Assays
}

