% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage.R
\name{Stage}
\alias{Stage}
\title{Stage a MultiAssayView by colnames, rownames, or assay}
\usage{
Stage(MultiAssay, identifier, method = character(), ...)
}
\arguments{
\item{MultiAssay}{A \code{\linkS4class{MultiAssayExperiment}}}

\item{method}{Prepare/Stage for subsetting using colnames, rownames or assays.}

\item{identifer}{Either a \code{character}, \code{numeric} or \code{logical} vector identifying targets}
}
\value{
A \code{\linkS4class{MultiAssayView}} class object for subsequent subsetting
}
\description{
Stage a MultiAssayView by colnames, rownames, or assay
}

