% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\docType{methods}
\name{subset,MultiAssayExperiment-method}
\alias{subset,MultiAssayExperiment-method}
\title{Subset MultiAssayExperiment object}
\usage{
\S4method{subset}{MultiAssayExperiment}(x, indicator, method = NULL,
  drop = TRUE, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{MultiAssayExperiment}} object}

\item{indicator}{A \code{logical} or \code{character} vector or
\code{GRanges} class object to use for subsetting}

\item{method}{A \code{character} vector of length one designating to subset
either by colnames, rownames, or assays}

\item{drop}{logical (default FALSE) whether to coerce lowest possible
dimension after subsetting}

\item{...}{Additional arguments to pass to
\code{\link[IRanges]{subsetByOverlaps}} when subsetting
by rownames}
}
\value{
A subsetted \link{MultiAssayExperiment} class object
}
\description{
The \code{subset} function allows for searching through colnames and
rownames to obtain the requested subset of data or experiments (assays).
}

