% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-methods.R
\name{subsetByAssay}
\alias{subsetByAssay}
\title{Subset MultiAssayExperiment object by Assay type}
\usage{
subsetByAssay(x, y)
}
\arguments{
\item{x}{A \code{\link{MultiAssayExperiment}} object}

\item{y}{Either a \code{numeric}, \code{character} or
\code{logical} object indicating what assay(s) to select}
}
\value{
A \code{\link{MultiAssayExperiment}} object
}
\description{
Select which assay(s) to obtain from available datasets
}

