% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetByAssay.R
\name{subsetByAssay}
\alias{subsetByAssay}
\title{Subset MultiAssayExperiment object by Assay type}
\usage{
subsetByAssay(MultiAssay, assayName, drop = FALSE)
}
\arguments{
\item{MultiAssay}{A \code{\link{MultiAssayExperiment}} object}

\item{assayName}{Either a numeric, character or logical vector indicating what assay(s) to select}

\item{drop}{logical Indicates whether to drop the unmatched assays from the object}
}
\value{
A \code{\link{MultiAssayExperiment}} object or other if drop paramater is set to TRUE
}
\description{
Select which assay(s) to obtain from available datasets,
fuzzy matching is possible but exact names recommended
}

