% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-methods.R
\docType{methods}
\name{subsetByAssay}
\alias{subsetByAssay}
\alias{subsetByAssay,MultiAssayExperiment-method}
\title{Subset \code{MultiAssayExperiment} object by Assay type}
\usage{
subsetByAssay(x, y)

\S4method{subsetByAssay}{MultiAssayExperiment}(x, y)
}
\arguments{
\item{x}{A \code{\link{MultiAssayExperiment}} object}

\item{y}{Either a \code{numeric}, \code{character} or
\code{logical} object indicating what assay(s) to select}
}
\value{
A \code{\link{MultiAssayExperiment}} object
}
\description{
Select which assay(s) to obtain from available datasets
}
\section{Methods (by class)}{
\itemize{
\item \code{MultiAssayExperiment}: Use either a \code{numeric}, \code{logical},
or \code{character} vector to subset assays in a
\code{MultiAssayExperiment}
}}

\examples{
## Load a MultiAssayExperiment example
example("MultiAssayExperiment")

## Using experiment names
subsetByAssay(myMultiAssayExperiment, "Affy")

## Using numeric indicators
subsetByAssay(myMultiAssayExperiment, 1:2)

## Using a logical vector
subsetByAssay(myMultiAssayExperiment, c(TRUE, FALSE, TRUE))

}
\seealso{
`subset,MultiAssayExperiment-method`
}
