% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetByColumn.R
\name{subsetByColumn}
\alias{subsetByColumn}
\title{Subset MultiAssayExperiment object
\code{subsetByColumn} returns a subsetted \code{\linkS4class{MultiAssayExperiment}} object}
\usage{
subsetByColumn(MultiAssay, identifier, drop = FALSE)
}
\arguments{
\item{MultiAssay}{A \code{\link{MultiAssayExperiment}} object}

\item{identifier}{A \linkS4class{MultiAssayView} class object to be used for subsetting}
}
\description{
Subset MultiAssayExperiment object
\code{subsetByColumn} returns a subsetted \code{\linkS4class{MultiAssayExperiment}} object
}

