% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocMultiAssay-methods.R
\name{subsetByColumn}
\alias{subsetByColumn}
\title{Subset MultiAssayExperiment object}
\usage{
subsetByColumn(x, y)
}
\arguments{
\item{x}{A \code{\link{MultiAssayExperiment}} object}

\item{y}{Either a \code{numeric}, \code{character} or
\code{logical} object indicating what rownames in the pData to select
for subsetting}
}
\value{
A \code{\link{MultiAssayExperiment}} object
}
\description{
\code{subsetByColumn} returns a subsetted 
\code{\linkS4class{MultiAssayExperiment}} object
}

