% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocMultiAssay-methods.R
\docType{methods}
\name{subsetByColumn}
\alias{subsetByColumn}
\alias{subsetByColumn,MultiAssayExperiment,ANY-method}
\alias{subsetByColumn,MultiAssayExperiment,character-method}
\title{Subset MultiAssayExperiment object}
\usage{
subsetByColumn(x, y)

\S4method{subsetByColumn}{MultiAssayExperiment,ANY}(x, y)

\S4method{subsetByColumn}{MultiAssayExperiment,character}(x, y)
}
\arguments{
\item{x}{A \code{\link{MultiAssayExperiment}} object}

\item{y}{Either a \code{numeric}, \code{character} or
\code{logical} object indicating what rownames in the pData to select
for subsetting}
}
\value{
A \code{\link{MultiAssayExperiment}} object
}
\description{
\code{subsetByColumn} returns a subsetted 
\code{\linkS4class{MultiAssayExperiment}} object
}
\section{Methods (by class)}{
\itemize{
\item \code{x = MultiAssayExperiment,y = ANY}: Either a \code{numeric} or \code{logical} vector
to apply a column subset of a \code{MultiAssayExperiment} object

\item \code{x = MultiAssayExperiment,y = character}: Use a character vector for subsetting column 
names
}}

