% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetByColumn.R
\name{subsetByColumn}
\alias{subsetByColumn}
\title{Subset MultiAssayExperiment object}
\usage{
subsetByColumn(MultiAssayExperiment, colIndicator)
}
\arguments{
\item{MultiAssayExperiment}{A \code{\link{MultiAssayExperiment}} object}

\item{colIndicator}{A \linkS4class{MultiAssayView} class object to be used 
for subsetting}
}
\value{
A \code{\link{MultiAssayExperiment}} object
}
\description{
\code{subsetByColumn} returns a subsetted 
\code{\linkS4class{MultiAssayExperiment}} object
}

