% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-methods.R
\docType{methods}
\name{subsetByColumn}
\alias{subsetByColumn}
\alias{subsetByColumn,MultiAssayExperiment,list-method}
\alias{subsetByColumn,MultiAssayExperiment,List-method}
\title{Subset \code{MultiAssayExperiment} object}
\usage{
subsetByColumn(x, y)

\S4method{subsetByColumn}{MultiAssayExperiment,list}(x, y)

\S4method{subsetByColumn}{MultiAssayExperiment,List}(x, y)
}
\arguments{
\item{x}{A \code{\link{MultiAssayExperiment}} object}

\item{y}{Either a \code{numeric}, \code{character} or
\code{logical} object indicating what rownames in the colData to select
for subsetting}
}
\value{
A \code{\link{MultiAssayExperiment}} object
}
\description{
\code{subsetByColumn} returns a subsetted
\code{\linkS4class{MultiAssayExperiment}} object
}
\section{Methods (by class)}{
\itemize{
\item \code{x = MultiAssayExperiment,y = list}: Use a \code{list} to subset by
colname in a \code{MultiAssayExperiment}

\item \code{x = MultiAssayExperiment,y = List}: Use an S4 \code{List} to subset
a \code{MultiAssayExperiment}. The order of the subsetting
elements in this \code{List} must match that of the
\code{ExperimentList} in the \code{MultiAssayExperiment}.
}}

\examples{
## Load a MultiAssayExperiment example
example("MultiAssayExperiment")

subsetByColumn(myMultiAssayExperiment, list(Affy = 1:2,
                Methyl450k = c(3,5,2), RNASeqGene = 2:4, CNVgistic = 1))

subsetWith <- IRanges::mendoapply(`[`, colnames(myMultiAssayExperiment),
                        MoreArgs = list(1:2))
subsetByColumn(myMultiAssayExperiment, subsetWith)

}
