% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetByFeature.R
\name{subsetByFeature}
\alias{subsetByFeature}
\title{Subset MultiAssayExperiment object by Feature}
\usage{
subsetByFeature(MultiAssay, featureIndicator, drop = FALSE)
}
\arguments{
\item{MultiAssay}{A \code{\link{MultiAssayExperiment}} object}

\item{featureIndicator}{Either a \code{character} vector or \code{\linkS4class{MultiAssayView}} class object containing feature names}
}
\value{
A \code{\link{MultiAssayExperiment}} object
}
\description{
Subset MultiAssayExperiment class by provided feature names
}

