% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetByRow.R
\name{subsetByRow}
\alias{subsetByRow}
\title{Subset MultiAssayExperiment object by Feature}
\usage{
subsetByRow(MultiAssayExperiment, rowIndicator, ...)
}
\arguments{
\item{MultiAssayExperiment}{A \code{\link{MultiAssayExperiment}} object}

\item{rowIndicator}{A \code{character} vector class object containing
feature names}

\item{...}{Additional arguments to pass to low level subsetting function}
}
\value{
A \code{\link{MultiAssayExperiment}} object
}
\description{
Subset MultiAssayExperiment class by provided feature names
}

