% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocMultiAssay-methods.R
\name{subsetByRow}
\alias{subsetByRow}
\title{Subset MultiAssayExperiment object by Feature}
\usage{
subsetByRow(x, y, ...)
}
\arguments{
\item{x}{A \code{\link{MultiAssayExperiment}} object}

\item{y}{A \code{character} vector or \code{GRanges} class object
containing feature names or ranges}

\item{...}{Additional arguments to pass to low level subsetting function 
primarily when using a \code{GRanges} object for subsetting
(via \code{getHits})}
}
\value{
A \code{\link{MultiAssayExperiment}} object
}
\description{
Subset \code{MultiAssayExperiment} class by provided feature names or a 
\code{GRanges} object
}
\seealso{
\code{\link{getHits}}
}

