% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-methods.R
\docType{methods}
\name{subsetByRow}
\alias{subsetByRow}
\alias{subsetByRow,MultiAssayExperiment,ANY-method}
\alias{subsetByRow,MultiAssayExperiment,list-method}
\alias{subsetByRow,MultiAssayExperiment,List-method}
\title{Subset \code{MultiAssayExperiment} object by Feature}
\usage{
subsetByRow(x, y, ...)

\S4method{subsetByRow}{MultiAssayExperiment,ANY}(x, y, ...)

\S4method{subsetByRow}{MultiAssayExperiment,list}(x, y)

\S4method{subsetByRow}{MultiAssayExperiment,List}(x, y)
}
\arguments{
\item{x}{A \code{\link{MultiAssayExperiment}} object}

\item{y}{A \code{character} vector or \code{GRanges} class object
containing feature names or ranges}

\item{...}{Additional arguments to pass to low level subsetting function
primarily when using a \code{GRanges} object for subsetting
(via \code{getHits})}
}
\value{
A \code{\link{MultiAssayExperiment}} object
}
\description{
Subset a \code{MultiAssayExperiment} class by provided feature names or a
\code{GRanges} object
}
\section{Methods (by class)}{
\itemize{
\item \code{x = MultiAssayExperiment,y = ANY}: Subset a
\code{MultiAssayExperiment} with either a
\code{numeric} or \code{logical} vector

\item \code{x = MultiAssayExperiment,y = list}: Use a list of equal length as
the \code{ExperimentList} to subset. The order of
the subsetting elements in this list must match that of the
\code{ExperimentList} in the \code{MultiAssayExperiment}.

\item \code{x = MultiAssayExperiment,y = List}: Use an S4 \code{List} to subset
a \code{MultiAssayExperiment}. The order of the subsetting elements in
this \code{List} must match that of the \code{ExperimentList} in the
\code{MultiAssayExperiment}.
}}

\examples{
## Load a MultiAssayExperiment example
example("MultiAssayExperiment")

## Use a GRanges object to subset rows where ranged data present
egr <- GRanges(seqnames = "chr1", IRanges(start = 1, end = 3), strand = "-")
subsetByRow(myMultiAssayExperiment, egr)

## Use a logical vector (recycling used)
subsetByRow(myMultiAssayExperiment, c(TRUE, FALSE))

## Use a character vector
subsetByRow(myMultiAssayExperiment, "ENST00000355076")

}
