% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayExperiment-methods.R
\docType{methods}
\name{subsetByRow}
\alias{subsetByRow}
\alias{subsetByRow,MultiAssayExperiment,ANY-method}
\alias{subsetByRow,MultiAssayExperiment,GRanges-method}
\alias{subsetByRow,MultiAssayExperiment,GRangesORcharacter-method}
\alias{subsetByRow,MultiAssayExperiment,List-method}
\alias{subsetByRow,MultiAssayExperiment,list-method}
\alias{subsetByRow,MultiAssayExperiment,logical-method}
\title{Subset MultiAssayExperiment object by Feature}
\usage{
subsetByRow(x, y, ...)

\S4method{subsetByRow}{MultiAssayExperiment,GRangesORcharacter}(x, y, ...)

\S4method{subsetByRow}{MultiAssayExperiment,GRanges}(x, y, ...)

\S4method{subsetByRow}{MultiAssayExperiment,logical}(x, y)

\S4method{subsetByRow}{MultiAssayExperiment,ANY}(x, y)

\S4method{subsetByRow}{MultiAssayExperiment,list}(x, y)

\S4method{subsetByRow}{MultiAssayExperiment,List}(x, y)
}
\arguments{
\item{x}{A \code{\link{MultiAssayExperiment}} object}

\item{y}{A \code{character} vector or \code{GRanges} class object
containing feature names or ranges}

\item{...}{Additional arguments to pass to low level subsetting function 
primarily when using a \code{GRanges} object for subsetting
(via \code{getHits})}
}
\value{
A \code{\link{MultiAssayExperiment}} object
}
\description{
Subset \code{MultiAssayExperiment} class by provided feature names or a 
\code{GRanges} object
}
\section{Methods (by class)}{
\itemize{
\item \code{x = MultiAssayExperiment,y = GRangesORcharacter}: Use either a GRanges or character to select the
rows for which to subset for

\item \code{x = MultiAssayExperiment,y = GRanges}: Subset MultiAssayExperiment with
GRanges object

\item \code{x = MultiAssayExperiment,y = logical}: Use a logical vector to select rows of a
MultiAssayExperiment

\item \code{x = MultiAssayExperiment,y = ANY}: Subset a MultiAssayExperiment with either a
numeric or logical vector

\item \code{x = MultiAssayExperiment,y = list}: Use a list of equal length as the Elist to subset

\item \code{x = MultiAssayExperiment,y = List}: Use an S4 List to subset a MultiAssayExperiment
}}
\seealso{
\code{\link{getHits}}
}

