% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subsetBySample.R
\name{subsetBySample}
\alias{subsetBySample}
\title{Subset MultiAssayExperiment object
\code{subsetBySample} returns a subsetted \code{\linkS4class{MultiAssayExperiment}} object}
\usage{
subsetBySample(MultiAssay, logicID)
}
\arguments{
\item{MultiAssay}{A \code{\linkS4class{MultiAssayExperiment}} object}

\item{logicID}{A list of logical indicators to be used for subsetting}
}
\description{
Subset MultiAssayExperiment object
\code{subsetBySample} returns a subsetted \code{\linkS4class{MultiAssayExperiment}} object
}

