% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biocMultiAssay-methods.R
\docType{methods}
\name{subsetFeature}
\alias{subsetFeature}
\alias{subsetFeature,ExpressionSet-method}
\alias{subsetFeature,SummarizedExperiment-method}
\alias{subsetFeature,matrix-method}
\title{Subset by Feature method}
\usage{
subsetFeature(x, j, ...)

\S4method{subsetFeature}{matrix}(x, j)

\S4method{subsetFeature}{ExpressionSet}(x, j)

\S4method{subsetFeature}{SummarizedExperiment}(x, j)
}
\arguments{
\item{x}{Either an \code{\linkS4class{ExpressionSet}}, \code{\linkS4class{GRangesList}}, \code{\linkS4class{SummarizedExperiment}} or \code{matrix} class object}

\item{j}{Either a \code{"numeric"} or \code{"character"} vector class for subsetting}

\item{...}{Additional arguments to pass}
}
\value{
Returnss a subsetted \code{\linkS4class{MultiAssayExperiment}} object
}
\description{
Subset by Feature method
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: 

\item \code{ExpressionSet}: 

\item \code{SummarizedExperiment}: 
}}

