% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biocMultiAssay-classes.R
\name{subsetSample}
\alias{subsetSample}
\title{Subset by Sample method for eSet, SummarizedExperiment, matrix, and GRangesList}
\usage{
subsetSample(x, j, ...)
}
\arguments{
\item{x}{Either an \code{\linkS4class{ExpressionSet}}, \code{\linkS4class{GRangesList}}, \code{\linkS4class{SummarizedExperiment}} or \code{matrix} class object}

\item{j}{Either a \code{"numeric"} or \code{"character"} vector class for subsetting}

\item{...}{Additional arguments to pass}
}
\value{
Returns a subsetted \code{\linkS4class{MultiAssayExperiment}} object
}
\description{
Subset by Sample method for eSet, SummarizedExperiment, matrix, and GRangesList
}

