% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toListMap.R
\name{toListMap}
\alias{toListMap}
\title{Convert data.frame map to list map
\code{toListMap} returns a \code{list}}
\usage{
toListMap(mapdf, assayCol = NULL)
}
\arguments{
\item{assayCol}{A character vector of length one indicating the assay names column}

\item{dfmap}{A data frame object with identifiers in the first column}
}
\description{
Convert data.frame map to list map
\code{toListMap} returns a \code{list}
}

